Extracted classes from Rewriter.cs.
