+++ Foxtrot/Driver/Rewriting/RewriteHelper.cs
@@ -12,14 +12,10 @@
 // 
 // THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Text;
 using System.Diagnostics; // needed for Debug.Assert (etc.)
 using System.Runtime.Serialization; // needed for defining exception .ctors
 using System.Compiler;
-using System.IO;
 using System.Diagnostics.Contracts;
 
 namespace Microsoft.Contracts.Foxtrot
