+++ Foxtrot/Foxtrot/Rewriter.cs
@@ -15,2762 +15,12 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics; // needed for Debug.Assert (etc.)
-using System.Runtime.Serialization; // needed for defining exception .ctors
+// needed for defining exception .ctors
 using System.Compiler;
-using System.IO;
 using System.Diagnostics.Contracts;
-using System.Diagnostics.CodeAnalysis;
 
 namespace Microsoft.Contracts.Foxtrot
 {
-    [Flags]
-    internal enum RuntimeContractEmitFlags
-    {
-        None = 0,
-        LegacyRequires = 0x0001,
-        RequiresWithException = 0x0002,
-        Requires = 0x0004,
-        Ensures = 0x0008,
-        Invariants = 0x0010,
-        Asserts = 0x0020,
-        Assumes = 0x0040,
-        AsyncEnsures = 0x0080,
-        ThrowOnFailure = 0x1000,
-        StandardMode = 0x2000,
-        InheritContracts = 0x4000,
-
-        /// <summary>
-        /// Takes precedence over individual bits (like a mask)
-        /// </summary>
-        NoChecking = 0x8000,
-    }
-
-    internal static class FlagExtensions
-    {
-        public static bool Emit(this RuntimeContractEmitFlags have, RuntimeContractEmitFlags want)
-        {
-            return ((have & RuntimeContractEmitFlags.NoChecking) == 0) && (have & want) != 0;
-        }
-    }
-
-    internal sealed class ReplaceReturns : StandardVisitor
-    {
-        private readonly Local result;
-        private readonly Block newExit;
-
-        private int returnCount;
-
-        private readonly bool leaveExceptionBody;
-        private SourceContext lastReturnSourceContext;
-
-        public ReplaceReturns(Local r, Block b, bool leaveExceptionBody = false)
-        {
-            result = r;
-            newExit = b;
-            this.leaveExceptionBody = leaveExceptionBody;
-        }
-
-        //public override Block VisitBlock(Block block) {
-        //  if(block.Statements != null && block.Statements.Count == 1) {
-        //    Return r = block.Statements[0] as Return;
-        //    if(r != null) {
-        //      Statement s = this.VisitReturn(r);
-        //      Block retBlock = s as Block;
-        //      if(retBlock != null) {
-        //        block.Statements = retBlock.Statements;
-        //        return block;
-        //      } else {
-        //        return base.VisitBlock(block);
-        //      }
-        //    } else {
-        //      return base.VisitBlock(block);
-        //    }
-        //  } else {
-        //    return base.VisitBlock(block);
-        //  }
-        //}
-
-        public override Statement VisitReturn(Return Return)
-        {
-            if (Return == null)
-            {
-                return null;
-            }
-
-            returnCount++;
-            this.lastReturnSourceContext = Return.SourceContext;
-
-            StatementList stmts = new StatementList();
-
-            Return.Expression = this.VisitExpression(Return.Expression);
-
-            if (Return.Expression != null)
-            {
-                MethodCall mc = Return.Expression as MethodCall;
-                if (mc != null && mc.IsTailCall)
-                {
-                    mc.IsTailCall = false;
-                }
-
-                var assgnmt = new AssignmentStatement(result, Return.Expression);
-
-                assgnmt.SourceContext = Return.SourceContext;
-                stmts.Add(assgnmt);
-            }
-
-            // the branch is a "leave" out of the try block that the body will be
-            // in.
-            var branch = new Branch(null, newExit, false, false, this.leaveExceptionBody);
-            branch.SourceContext = Return.SourceContext;
-
-            stmts.Add(branch);
-
-            return new Block(stmts);
-        }
-
-        public SourceContext LastReturnSourceContext
-        {
-            get { return this.lastReturnSourceContext; }
-        }
-    }
-
-    internal sealed class ReplaceResult : StandardVisitor
-    {
-        /// <summary>
-        /// Result starts out as a Local that is created by the client of this class. But if during
-        /// the replacement of Result&lt;T&gt;(), an occurrence is found within an anonymous delegate,
-        /// then the local is replaced with a member binding.
-        /// 
-        /// If the delegate is within a closure class, then the member binding is "this.f", where
-        /// f is a new field that is added to the top-level closure class and "this" will be the
-        /// type of the top-level closure class.
-        /// 
-        /// If the delegate is a static method added to the class itself, then the member binding
-        /// is "null.f" where f is a new static field that is added to the class containing the method
-        /// that contained the contract in which this occurrence of Result was found.
-        /// 
-        /// When this happens, a client should use the value of ReplaceResult.Result
-        /// to assign the return value of the method to.
-        /// 
-        /// Note: The client should assign the value that the method returns to *both* the local
-        /// originally passed in to the constructor of this visitor *and* to the member binding
-        /// (if Result is modified because there is an occurrence of Result&lt;T&gt()
-        /// in a delegate) because if any occurrences are found before visiting the closure,
-        /// those occurrences will be replaced by the local and not the member binding!
-        /// 
-        /// When Result is found *not* within a delegate, then it is always replaced with
-        /// the original local that the client handed to the constructor.
-        private readonly Local originalLocalForResult;
-
-        /// if non-null, used to store result in declaring type.
-        /// NOTE: the field type is the type of the
-        /// result in the instantiated context. For generic closure classes (which are generated by the 
-        /// rewriter during copying of OOB contracts), the closure potentially expects a generic form for
-        /// the result value. We thus have to generate access to the result by casting (box InstanceType, unbox GenericType).
-        /// </summary>
-        private Field topLevelStaticResultField;
-
-        /// <summary>
-        /// if non-null, used to store result in top-level closure instance. 
-        /// NOTE: the field type is the type of the
-        /// result in the instantiated context. For generic closure classes (which are generated by the 
-        /// rewriter during copying of OOB contracts), the closure potentially expects a generic form for
-        /// the result value. We thus have to generate access to the result by casting (box InstanceType, unbox GenericType).
-        /// </summary>
-        private Field topLevelClosureResultField;
-
-        public IEnumerable<MemberBinding> NecessaryResultInitialization(Dictionary<TypeNode, Local> closureLocals)
-        {
-            if (topLevelStaticResultField != null)
-            {
-                yield return new MemberBinding(null, topLevelStaticResultField);
-            }
-
-            if (topLevelClosureResultField != null)
-            {
-                // note: this field is the field in the generic context of the closure. For the method to initialize this
-                // we need the instantiated form, which we remember in topLevelClosureClassInstance
-                Local local;
-                if (closureLocals.TryGetValue(topLevelClosureClassInstance, out local))
-                {
-                    yield return
-                        new MemberBinding(local,
-                            GetInstanceField(this.originalLocalForResult.Type, topLevelClosureResultField,
-                                topLevelClosureClassInstance));
-                }
-                else
-                {
-                    var access = new This(topLevelClosureClassInstance);
-                    yield return
-                        new MemberBinding(access,
-                            GetInstanceField(this.originalLocalForResult.Type, topLevelClosureResultField,
-                                topLevelClosureClassInstance));
-                }
-            }
-        }
-
-        public IEnumerable<MemberBinding> NecessaryResultInitializationAsync(Dictionary<Local, MemberBinding> closureLocals)
-        {
-            if (topLevelStaticResultField != null)
-            {
-                yield return new MemberBinding(null, topLevelStaticResultField);
-            }
-
-            if (topLevelClosureResultField != null)
-            {
-                // note: this field is the field in the generic context of the closure. For the method to initialize this
-                // we need the instantiated form, which we remember in topLevelClosureClassInstance
-                MemberBinding mb = null;
-                foreach (var pair in closureLocals)
-                {
-                    var keyType = HelperMethods.Unspecialize(pair.Key.Type);
-                    if (keyType == topLevelClosureClassDefinition)
-                    {
-                        mb = pair.Value;
-                        yield return
-                            new MemberBinding(mb,
-                                GetInstanceField(this.originalLocalForResult.Type, topLevelClosureResultField,
-                                    topLevelClosureClassInstance));
-                        break;
-                    }
-                }
-
-                Debug.Assert(mb != null, "Should have found access");
-            }
-        }
-
-        private readonly Module assemblyBeingRewritten;
-        private Method currentClosureMethod; // definition
-        private Expression currentAccessToTopLevelClosure;
-
-        private TypeNode topLevelClosureClassDefinition;
-        private TypeNode topLevelClosureClassInstance;
-        private TypeNode currentClosureClassInstance;
-        private readonly TypeNodeList topLevelMethodFormals;
-
-        /// <summary>
-        /// Used when the method with the closure is generic and the field ends up on the corresponding generic closure class
-        /// </summary>
-        private TypeNode properlyInstantiatedFieldType;
-
-        private int delegateNestingLevel;
-        private readonly TypeNode declaringType; // needed to copy anonymous delegates into
-
-        public ReplaceResult(Method containingMethod, Local r, Module assemblyBeingRewritten)
-        {
-            Contract.Requires(containingMethod != null);
-
-            this.assemblyBeingRewritten = assemblyBeingRewritten;
-            this.declaringType = containingMethod.DeclaringType;
-            this.topLevelMethodFormals = containingMethod.TemplateParameters;
-            this.originalLocalForResult = r;
-            this.delegateNestingLevel = 0;
-        }
-
-        /// <summary>
-        /// This property would be true, if Contract.Result&lt;T&gt;() was captured in the static context.
-        /// For instance, following code will lead to this situation: 
-        ///         public string Method(params string[] strings)
-        ///    {
-        ///        Contract.Ensures(Contract.ForAll(strings, s => Contract.Result&lt;string&gt;() == s));
-        ///        return "42";
-        ///    }
-        ///
-        /// In this case, caller of this code can decide to emit a warning, because currently resultin 
-        /// </summary>
-        public bool ContractResultWasCapturedInStaticContext { get; private set; }
-
-        public override Expression VisitReturnValue(ReturnValue returnValue)
-        {
-            if (this.delegateNestingLevel == 0)
-            {
-                // not inside a closure method
-                return this.originalLocalForResult;
-            }
-
-            if (this.currentClosureMethod.IsStatic || IsRoslynBasedStaticClosure())
-            {
-                // This is a hack and we should notify about it!
-                ContractResultWasCapturedInStaticContext = true;
-
-                // static closure: no place to store result. Current hack is to use a static field of the 
-                // declaring type. However, if we have a static closure inside a non-static closure, this
-                // breaks down, as we support only one kind of storage for return values.
-                var field = GetReturnValueClosureField(this.declaringType, this.originalLocalForResult.Type,
-                    FieldFlags.CompilerControlled | FieldFlags.Private | FieldFlags.Static,
-                    originalLocalForResult.UniqueKey);
-
-                Contract.Assume(returnValue != null);
-
-                return CreateProperResultAccess(returnValue, null, field);
-            }
-
-            Debug.Assert(this.currentAccessToTopLevelClosure != null);
-            Debug.Assert(this.topLevelClosureClassDefinition != null);
-
-            {
-                // Return an expression that is the value of the field defined in the
-                // top-level closure class to hold the method's return value.
-                // This will be this.up.Result where "up" is the field C#
-                // generated to point to the instance of the top-level closure class.
-                // "Result" is the field defined in this visitor's VisitConstruct when
-                // it finds a reference to a anonymous delegate.
-                var field = GetReturnValueClosureField(this.topLevelClosureClassDefinition,
-                    this.properlyInstantiatedFieldType, FieldFlags.CompilerControlled | FieldFlags.Assembly,
-                    this.topLevelClosureClassDefinition.UniqueKey);
-
-                Contract.Assume(returnValue != null);
-
-                return CreateProperResultAccess(returnValue, this.currentAccessToTopLevelClosure, field);
-            }
-        }
-
-        /// <summary>
-        /// Roslyn-based compiler changed the pattern for caching static (i.e. non-capturing) lambdas.
-        /// This method returns true if currentClosureClassInstance is a static closure class generated by the Roslyn-based compiler.
-        /// </summary>
-        private bool IsRoslynBasedStaticClosure()
-        {
-            return currentClosureClassInstance.IsRoslynBasedStaticClosure();
-        }
-
-        private static Expression CreateProperResultAccess(ReturnValue returnValue, Expression closureObject, Field resultField)
-        {
-            Contract.Requires(returnValue != null);
-            Contract.Requires(resultField != null);
-
-            var fieldAccess = new MemberBinding(closureObject, resultField);
-
-            if (resultField.Type != returnValue.Type)
-            {
-                // must cast to generic type expected in this context (box instance unbox.any Generic)
-                return
-                    new BinaryExpression(
-                        new BinaryExpression(fieldAccess, new Literal(resultField.Type), NodeType.Box),
-                        new Literal(returnValue.Type), NodeType.UnboxAny);
-            }
-            
-            return fieldAccess;
-        }
-
-        /// <summary>
-        /// If there is an anonymous delegate within a postcondition, then there
-        /// will be a call to a delegate constructor.
-        /// That call looks like "d..ctor(o,m)" where d is the type of the delegate.
-        /// There are two cases depending on whether the anonymous delegate captured
-        /// anything. In both cases, m is the method implementing the anonymous delegate.
-        /// (1) It does capture something. Then o is the instance of the closure class
-        /// implementing the delegate, and m is an instance method in the closure
-        /// class.
-        /// (2) It does *not* capture anything. Then o is the literal for null and
-        /// m is a static method that was added directly to the class.
-        /// 
-        /// This method will cause the method (i.e., m) to be visited to collect any
-        /// Result&lt;T&gt;() expressions that occur in it.
-        /// </summary>
-        /// <param name="cons">The AST representing the call to the constructor
-        /// of the delegate type.</param>
-        /// <returns>Whatever the base visitor returns</returns>
-        public override Expression VisitConstruct(Construct cons)
-        {
-            if (cons.Type is DelegateNode)
-            {
-                UnaryExpression ue = cons.Operands[1] as UnaryExpression;
-                if (ue == null) goto JustVisit;
-
-                MemberBinding mb = ue.Operand as MemberBinding;
-                if (mb == null) goto JustVisit;
-
-                Method m = mb.BoundMember as Method;
-                if (!HelperMethods.IsCompilerGenerated(m)) goto JustVisit;
-
-                Contract.Assume(m != null);
-
-                m = Definition(m);
-                this.delegateNestingLevel++;
-
-                TypeNode savedClosureClass = this.currentClosureClassInstance;
-                Method savedClosureMethod = this.currentClosureMethod;
-                Expression savedCurrentAccessToTopLevelClosure = this.currentAccessToTopLevelClosure;
-
-                try
-                {
-                    this.currentClosureMethod = m;
-
-                    if (m.IsStatic)
-                    {
-                        this.currentClosureClassInstance = null; // no closure object
-                    }
-                    else
-                    {
-                        this.currentClosureClassInstance = cons.Operands[0].Type;
-                        if (savedClosureClass == null)
-                        {
-                            // Then this is the top-level closure class.
-                            this.topLevelClosureClassInstance = this.currentClosureClassInstance;
-                            this.topLevelClosureClassDefinition = Definition(this.topLevelClosureClassInstance);
-
-                            this.currentAccessToTopLevelClosure = new This(this.topLevelClosureClassDefinition);
-                            this.properlyInstantiatedFieldType = this.originalLocalForResult.Type;
-
-                            if (this.topLevelMethodFormals != null)
-                            {
-                                Contract.Assume(this.topLevelClosureClassDefinition.IsGeneric);
-                                Contract.Assume(topLevelClosureClassDefinition.TemplateParameters.Count >=
-                                                this.topLevelMethodFormals.Count);
-
-                                // replace method type parameters in result properly with last n corresponding type parameters of closure class
-                                TypeNodeList closureFormals = topLevelClosureClassDefinition.TemplateParameters;
-                                if (closureFormals.Count > this.topLevelMethodFormals.Count)
-                                {
-                                    int offset = closureFormals.Count - this.topLevelMethodFormals.Count;
-                                    closureFormals = new TypeNodeList(this.topLevelMethodFormals.Count);
-                                    for (int i = 0; i < this.topLevelMethodFormals.Count; i++)
-                                    {
-                                        closureFormals.Add(topLevelClosureClassDefinition.TemplateParameters[i + offset]);
-                                    }
-                                }
-
-                                Duplicator dup = new Duplicator(this.declaringType.DeclaringModule, this.declaringType);
-
-                                Specializer spec = new Specializer(this.declaringType.DeclaringModule,
-                                    topLevelMethodFormals, closureFormals);
-
-                                var type = dup.VisitTypeReference(this.originalLocalForResult.Type);
-                                type = spec.VisitTypeReference(type);
-                                this.properlyInstantiatedFieldType = type;
-                            }
-                        }
-                        else
-                        {
-                            while (currentClosureClassInstance.Template != null)
-                                currentClosureClassInstance = currentClosureClassInstance.Template;
-
-                            // Find the field in this.closureClass that the C# compiler generated
-                            // to point to the top-level closure
-                            foreach (Member mem in this.currentClosureClassInstance.Members)
-                            {
-                                Field f = mem as Field;
-                                if (f == null) continue;
-                                if (f.Type == this.topLevelClosureClassDefinition)
-                                {
-                                    var consolidatedTemplateParams = this.currentClosureClassInstance.ConsolidatedTemplateParameters;
-
-                                    TypeNode thisType;
-                                    if (consolidatedTemplateParams != null && consolidatedTemplateParams.Count > 0)
-                                    {
-                                        thisType =
-                                            this.currentClosureClassInstance.GetGenericTemplateInstance(
-                                                this.assemblyBeingRewritten, consolidatedTemplateParams);
-                                    }
-                                    else
-                                    {
-                                        thisType = this.currentClosureClassInstance;
-                                    }
-
-                                    this.currentAccessToTopLevelClosure = new MemberBinding(new This(thisType), f);
-
-                                    break;
-                                }
-                            }
-                        }
-                    }
-
-                    this.VisitBlock(m.Body);
-                }
-                finally
-                {
-                    this.delegateNestingLevel--;
-                    this.currentClosureMethod = savedClosureMethod;
-                    this.currentClosureClassInstance = savedClosureClass;
-                    this.currentAccessToTopLevelClosure = savedCurrentAccessToTopLevelClosure;
-                }
-            }
-
-        JustVisit:
-            return base.VisitConstruct(cons);
-        }
-
-        private static Method Definition(Method m)
-        {
-            Contract.Requires(m != null);
-
-            while (m.Template != null) m = m.Template;
-            return m;
-        }
-
-        private static TypeNode Definition(TypeNode t)
-        {
-            Contract.Requires(t != null);
-
-            while (t.Template != null) t = t.Template;
-            return t;
-        }
-
-        private Field GetReturnValueClosureField(TypeNode declaringType, TypeNode resultType, FieldFlags flags, int uniqueKey)
-        {
-            Contract.Requires(declaringType != null);
-
-            Contract.Assume(declaringType.Template == null);
-            Identifier name = Identifier.For("_result" + uniqueKey.ToString()); // unique name for this field
-
-            Field f = declaringType.GetField(name);
-            if (f != null) return f;
-
-            f = new Field(declaringType,
-                null,
-                flags,
-                name,
-                resultType,
-                null);
-
-            declaringType.Members.Add(f);
-            // remember we added it so we can make it part of initializations
-            if (f.IsStatic)
-            {
-                topLevelStaticResultField = f;
-            }
-            else
-            {
-                topLevelClosureResultField = f;
-            }
-
-            return f;
-        }
-
-        private static Field GetInstanceField(TypeNode originalReturnType, Field possiblyGenericField, TypeNode instanceDeclaringType)
-        {
-            Contract.Requires(instanceDeclaringType != null);
-
-            if (instanceDeclaringType.Template == null) return possiblyGenericField;
-
-            var declaringTemplate = instanceDeclaringType;
-
-            while (declaringTemplate.Template != null)
-            {
-                declaringTemplate = declaringTemplate.Template;
-            }
-
-            Contract.Assume(declaringTemplate == possiblyGenericField.DeclaringType);
-
-            return Rewriter.GetFieldInstanceReference(possiblyGenericField, instanceDeclaringType);
-#if false
-      Field f = instanceDeclaringType.GetField(possiblyGenericField.Name);
-      if (f != null)
-      {
-        // already instantiated
-        return f;
-      }
-      // pseudo instance
-      Field instance = new Field(instanceDeclaringType, possiblyGenericField.Attributes, possiblyGenericField.Flags, possiblyGenericField.Name, originalReturnType, null);
-      instanceDeclaringType.Members.Add(instance);
-      return instance;
-#endif
-        }
-    }
-
-    internal sealed class CollectBoundVariables : Inspector
-    {
-        public readonly List<Parameter> FoundVariables = null;
-        public readonly List<Expression> FoundReferences = null;
-        private readonly List<Parameter> BoundVars;
-
-        public CollectBoundVariables(List<Parameter> boundVars)
-        {
-            Contract.Requires(boundVars != null);
-
-            this.FoundVariables = new List<Parameter>(boundVars.Count);
-            this.FoundReferences = new List<Expression>(boundVars.Count);
-            this.BoundVars = boundVars;
-        }
-
-        public override void VisitParameter(Parameter parameter)
-        {
-            if (parameter == null) return;
-            if (this.BoundVars.Contains(parameter) && !this.FoundVariables.Contains(parameter))
-            {
-                this.FoundVariables.Add(parameter);
-                this.FoundReferences.Add(parameter);
-            }
-        }
-
-        public override void VisitMemberBinding(MemberBinding memberBinding)
-        {
-            if (memberBinding == null) return;
-
-            if (memberBinding.TargetObject.NodeType == NodeType.This
-                || memberBinding.TargetObject.NodeType == NodeType.Local)
-            {
-                // search in list of parameters to see if any have the same name as the bound member
-                foreach (Parameter p in this.BoundVars)
-                {
-                    if (p.Name.UniqueIdKey == memberBinding.BoundMember.Name.UniqueIdKey)
-                    {
-                        if (!this.FoundVariables.Contains(p))
-                        {
-                            this.FoundVariables.Add(p);
-                            this.FoundReferences.Add(memberBinding);
-                        }
-                    }
-                }
-            }
-
-            base.VisitMemberBinding(memberBinding);
-        }
-    }
-
-    // FIXME! Need to do a sanity check that the argument to the Old method
-    // is a member binding (or at least something that is directly there and
-    // not something sitting on the stack as a result of some code sequence
-    // that this class doesn't capture and store in the OldExpressions list).
-    internal sealed class CollectOldExpressions : StandardVisitor
-    {
-        public Block PrestateValuesOfOldExpressions
-        {
-            get { return this.prestateValuesOfOldExpressions; }
-        }
-
-        private readonly Block prestateValuesOfOldExpressions;
-        private readonly Dictionary<TypeNode, Local> closureLocals;
-        private TypeNode topLevelClosureClass;
-        private TypeNode currentClosureClass;
-        private Field PointerToTopLevelClosureClass;
-
-        private readonly ContractNodes contractNodes;
-
-        private readonly List<MethodCall> stackOfMethods; // contains nested calls of ForAll and Exists methods
-
-        private readonly List<Parameter> stackOfBoundVariables;
-            // contains the parameters of the ForAll and Exists methods
-
-        private readonly Module module;
-        private readonly Method currentMethod;
-
-        private int counter;
-
-        public int Counter
-        {
-            get { return this.counter; }
-        }
-
-        public CollectOldExpressions(Module module, Method method, ContractNodes contractNodes, Dictionary<TypeNode, Local> closureLocals, 
-            int localCounterStart, Class initialClosureClass)
-            : this(module, method, contractNodes, closureLocals, localCounterStart)
-        {
-            this.topLevelClosureClass = initialClosureClass;
-            this.currentClosureClass = initialClosureClass;
-        }
-
-        public CollectOldExpressions(Module module, Method method, ContractNodes contractNodes, Dictionary<TypeNode, Local> closureLocals,
-            int localCounterStart)
-        {
-            this.contractNodes = contractNodes;
-            this.prestateValuesOfOldExpressions = new Block(new StatementList());
-            this.closureLocals = closureLocals;
-
-            this.stackOfMethods = new List<MethodCall>();
-            this.stackOfBoundVariables = new List<Parameter>();
-
-            this.module = module;
-            this.currentMethod = method;
-            this.counter = localCounterStart;
-        }
-
-        public override Expression VisitOldExpression(OldExpression oldExpression)
-        {
-            if (this.topLevelClosureClass != null)
-            {
-                // In Closure ==> Create a field
-
-                // Since we're within a closure, we can't create a local to hold the value of the old expression
-                // but instead have to create a field for it. That field can be a member of the top-level
-                // closure class since nothing mentioned in the old expression (except possibly for the
-                // bound variables of enclosing quantifications) should be anything captured from
-                // an inner anonymous delegate.
-
-                // BUT, first we have to know if the old expression depends on any of the bound
-                // variables of the closures in which it is located. If not, then we can implement
-                // it as a scalar and just generate the assignment "closure_class.field := e" for
-                // "Old(e)" to take a snapshot of e's value in the prestate. If it does depend on
-                // any of the bound variables, then we need to generate a set of for-loops that
-                // compute the indices and values of e for each tuple of indices so it can be retrieved
-                // (given the indices) in the post-state.
-                CollectBoundVariables cbv = new CollectBoundVariables(this.stackOfBoundVariables);
-                cbv.VisitExpression(oldExpression.expression);
-
-                SubstituteClosureClassWithinOldExpressions subst = new SubstituteClosureClassWithinOldExpressions(this.closureLocals);
-                Expression e = subst.VisitExpression(oldExpression.expression);
-                if (cbv.FoundVariables.Count == 0)
-                {
-                    // Use a scalar for the old variable
-
-                    Local closureLocal;
-                    if (!this.closureLocals.TryGetValue(this.topLevelClosureClass, out closureLocal))
-                    {
-                        Contract.Assume(false, "can't find closure local!");
-                    }
-
-                    // Define a scalar
-
-                    var clTemplate = HelperMethods.Unspecialize(this.topLevelClosureClass);
-                    Field f = new Field(clTemplate,
-                        null,
-                        FieldFlags.CompilerControlled | FieldFlags.Public,
-                        Identifier.For("_old" + oldExpression.expression.UniqueKey.ToString()),
-                        // unique name for this old expr.
-                        oldExpression.Type,
-                        null);
-
-                    clTemplate.Members.Add(f);
-
-                    // now produce properly instantiated field
-                    f = (Field) Rewriter.GetMemberInstanceReference(f, this.topLevelClosureClass);
-
-                    // Generate code to store value in prestate
-
-                    this.prestateValuesOfOldExpressions.Statements.Add(
-                        new AssignmentStatement(new MemberBinding(closureLocal, f), e));
-
-                    // Return expression to be used in poststate
-
-                    // Return an expression that will evaluate in the poststate to the value of the old
-                    // expression in the prestate. This will be this.up.f where "up" is the field C#
-                    // generated to point to the instance of the top-level closure class.
-                    if (this.PointerToTopLevelClosureClass == null)
-                    {
-                        // then the old expression occurs in the top-level closure class. Just return "this.f"
-                        // where "this" refers to the top-level closure class.
-                        return new MemberBinding(new This(this.currentClosureClass), f);
-                    }
-                    else
-                    {
-                        return new MemberBinding(
-                            new MemberBinding(new This(this.currentClosureClass), this.PointerToTopLevelClosureClass),
-                            f);
-                    }
-                }
-                else
-                {
-                    // the Old expression *does* depend upon at least one of the bound variable
-                    // in a ForAll or Exists expression
-
-                    // Use an indexed variable for the old variable
-
-                    TypeNode oldVariableTypeDomain;
-
-                    // Decide if domain is one-dimensional or not
-
-                    bool oneDimensional = cbv.FoundVariables.Count == 1 && cbv.FoundVariables[0].Type.IsValueType;
-                    if (oneDimensional)
-                    {
-                        // a one-dimensional old-expression can use the index variable directly
-                        oldVariableTypeDomain = cbv.FoundVariables[0].Type;
-                    }
-                    else
-                    {
-                        oldVariableTypeDomain = SystemTypes.GenericList.GetTemplateInstance(this.module,
-                            SystemTypes.Int32);
-                    }
-
-                    TypeNode oldVariableTypeRange = oldExpression.Type;
-                    TypeNode oldVariableType = SystemTypes.GenericDictionary.GetTemplateInstance(this.module,
-                        oldVariableTypeDomain, oldVariableTypeRange);
-                    
-                    Local closureLocal;
-                    if (!this.closureLocals.TryGetValue(this.topLevelClosureClass, out closureLocal))
-                    {
-                        Contract.Assume(false, "can't find closure local");
-                    }
-
-                    // Define an indexed variable
-
-                    var clTemplate = HelperMethods.Unspecialize(this.topLevelClosureClass);
-
-                    Field f = new Field(clTemplate,
-                        null,
-                        FieldFlags.CompilerControlled | FieldFlags.Assembly,
-                        // can't be private or protected because it needs to be accessed from inner (closure) classes that don't inherit from the class this field is added to.
-                        Identifier.For("_old" + oldExpression.expression.UniqueKey.ToString()),
-                        // unique name for this old expr.
-                        oldVariableType,
-                        null);
-
-                    clTemplate.Members.Add(f);
-                    
-                    // instantiate f
-                    f = (Field) Rewriter.GetMemberInstanceReference(f, closureLocal.Type);
-
-                    // Generate code to initialize the indexed variable
-
-                    Statement init = new AssignmentStatement(
-                        new MemberBinding(closureLocal, f),
-                        new Construct(new MemberBinding(null, oldVariableType.GetConstructor()), null));
-
-                    this.prestateValuesOfOldExpressions.Statements.Add(init);
-
-                    // Generate code to store values in prestate
-
-                    // Create assignment: this.closure.f[i,j,k,...] = e;
-
-                    Method setItem = oldVariableType.GetMethod(Identifier.For("set_Item"), oldVariableTypeDomain, oldVariableTypeRange);
-
-                    Expression index;
-                    if (oneDimensional)
-                    {
-                        index = cbv.FoundVariables[0];
-                    }
-                    else
-                    {
-                        //InstanceInitializer ctor =
-                        //  ContractNodes.TupleClass.GetConstructor(SystemTypes.Int32.GetArrayType(1));
-                        //Expression index = new Construct(new MemberBinding(null,ctor),new ExpressionList(
-                        index = Literal.Null;
-                    }
-
-                    MethodCall mc = new MethodCall(new MemberBinding(new MemberBinding(closureLocal, f), setItem),
-                        new ExpressionList(index, e));
-
-                    Statement stat = new ExpressionStatement(mc);
-
-                    List<Local> locals = new List<Local>(this.stackOfBoundVariables.Count);
-                    TrivialHashtable paramMap = new TrivialHashtable();
-
-                    // Generate a local for each bound variable to use in for-loop
-
-                    foreach (Variable v in this.stackOfBoundVariables)
-                    {
-                        Local l = new Local(Identifier.Empty, v.Type);
-                        paramMap[v.UniqueKey] = l;
-                        locals.Add(l);
-                    }
-
-                    // Substitute locals for bound variables in old expression *AND* in inner loop bounds
-
-                    SubstituteParameters sps = new SubstituteParameters(paramMap, this.stackOfBoundVariables);
-                    sps.Visit(stat);
-
-                    // Create nested for-loops around assignment
-
-                    // keep track of when the first variable is used (from innermost to outermost)
-                    // as soon as the first one is needed because the old expression depends on it,
-                    // then keep all enclosing loops. It would be possible to keep only those where
-                    // the necessary loops have loop bounds that depend on an enclosing loop, but I
-                    // haven't calculated that, so just keep them all. For instance, if the old expression
-                    // depends on j and the loops are "for i,0,n" and inside that "for j,0,i", then need
-                    // both loops. If the inner loop bounds were 0 and n, then wouldn't need the outer
-                    // loop.
-                    bool usedAVariable = false;
-
-                    for (int i = this.stackOfBoundVariables.Count - 1; 0 <= i; i--)
-                    {
-                        if (!usedAVariable
-                            && !cbv.FoundVariables.Contains(this.stackOfBoundVariables[i])) continue;
-                        usedAVariable = true;
-                        Expression lowerBound = new Duplicator(this.module, this.currentClosureClass).VisitExpression(
-                            this.stackOfMethods[i].Operands[0]);
-
-                        lowerBound = subst.VisitExpression(lowerBound);
-                        lowerBound = sps.VisitExpression(lowerBound);
-
-                        Expression upperBound = new Duplicator(this.module, this.currentClosureClass).VisitExpression(
-                            this.stackOfMethods[i].Operands[1]);
-
-                        upperBound = subst.VisitExpression(upperBound);
-                        upperBound = sps.VisitExpression(upperBound);
-
-                        stat = RewriteHelper.GenerateForLoop(locals[i], lowerBound, upperBound, stat);
-                    }
-
-                    this.prestateValuesOfOldExpressions.Statements.Add(stat);
-
-                    // Return expression to be used in poststate
-
-                    Method getItem = oldVariableType.GetMethod(Identifier.For("get_Item"), oldVariableTypeDomain);
-                    if (oneDimensional)
-                    {
-                        index = cbv.FoundReferences[0];
-                    }
-                    else
-                    {
-                        //InstanceInitializer ctor =
-                        //  ContractNodes.TupleClass.GetConstructor(SystemTypes.Int32.GetArrayType(1));
-                        //Expression index = new Construct(new MemberBinding(null,ctor),new ExpressionList(
-                        index = Literal.Null;
-                    }
-
-                    // Return an expression that will evaluate in the poststate to the value of the old
-                    // expression in the prestate. This will be this.up.f[i,j,k,...] where "up" is the field C#
-                    // generated to point to the instance of the top-level closure class.
-                    MemberBinding thisDotF;
-
-                    if (this.PointerToTopLevelClosureClass == null)
-                    {
-                        // then the old expression occurs in the top-level closure class. Just return "this.f"
-                        // where "this" refers to the top-level closure class.
-                        Contract.Assume(f != null);
-
-                        thisDotF = new MemberBinding(new This(clTemplate), HelperMethods.Unspecialize(f));
-                    }
-                    else
-                    {
-                        thisDotF = new MemberBinding(
-                            new MemberBinding(new This(clTemplate), this.PointerToTopLevelClosureClass),
-                            f);
-                    }
-
-                    return new MethodCall(new MemberBinding(thisDotF, getItem), new ExpressionList(index));
-                }
-            }
-            else
-            {
-                // Not in closure ==> Create a local variable
-
-                Local l = GetLocalForOldExpression(oldExpression);
-
-                // Make sure local can be seen in the debugger (for the entire method, unfortunately)
-
-                if (currentMethod.LocalList == null)
-                {
-                    currentMethod.LocalList = new LocalList();
-                }
-
-                currentMethod.LocalList.Add(l);
-                currentMethod.Body.HasLocals = true;
-
-                this.prestateValuesOfOldExpressions.Statements.Add(new AssignmentStatement(l, oldExpression.expression));
-
-                // Return an expression that will evaluate in the poststate to the value of the old
-                // expression in the prestate. When we're not in a closure, this is just the local
-                // itself.
-                return l;
-            }
-        }
-
-        private string BestGuessForLocalName(OldExpression oldExpression)
-        {
-            Contract.Requires(oldExpression != null);
-
-            MemberBinding mb = oldExpression.expression as MemberBinding;
-            if (mb != null)
-            {
-                return mb.BoundMember.Name.Name;
-            }
-
-            MethodCall mc = oldExpression.expression as MethodCall;
-            if (mc != null)
-            {
-                mb = mc.Callee as MemberBinding;
-                if (mb != null)
-                {
-                    Method calledMethod = mb.BoundMember as Method;
-                    if (calledMethod != null && calledMethod.IsPropertyGetter)
-                    {
-                        Property prop = calledMethod.DeclaringMember as Property;
-                        if (prop != null)
-                            return prop.Name.Name;
-                        
-                        return mb.BoundMember.Name.Name;
-                    }
-                    
-                    return mb.BoundMember.Name.Name;
-                }
-            }
-
-            Parameter p = oldExpression.expression as Parameter;
-            if (p != null)
-            {
-                return p.Name.Name;
-            }
-
-            // Didn't find something good to use for the name, so just make it unique.
-            var x = this.counter.ToString();
-            this.counter++;
-
-            return x;
-        }
-
-        private Local GetLocalForOldExpression(OldExpression oldExpression)
-        {
-            Contract.Requires(oldExpression != null);
-
-            string localName = BestGuessForLocalName(oldExpression);
-            return new Local(Identifier.For("Contract.Old(" + localName + ")"), oldExpression.Type);
-        }
-
-        public override Expression VisitMethodCall(MethodCall call)
-        {
-            if (call == null || call.Callee == null)
-            {
-                return call;
-            }
-
-            MemberBinding mb = call.Callee as MemberBinding;
-            if (mb == null)
-            {
-                return base.VisitMethodCall(call);
-            }
-
-            Method m = mb.BoundMember as Method;
-            if (m == null)
-            {
-                return base.VisitMethodCall(call);
-            }
-
-            // ForAll(...) or Exists(...)
-
-            if (this.contractNodes.IsForallMethod(m) || this.contractNodes.IsExistsMethod(m))
-            {
-                Parameter closureParameter = null;
-
-                // "ForAll(lb,ub, delegate(int i) { ... })" ==> "i" (or Exists)
-                ExpressionList operands = call.Operands;
-                
-                if (operands == null || operands.Count < 3) goto callBase;
-                
-                Construct construct = operands[2] as Construct;
-                if (construct == null)
-                {
-                    // then the predicate is just a static method because it didn't capture
-                    // and parameters and so was not a method on a closure class
-                    goto callBase;
-                }
-
-                ExpressionList operands2 = construct.Operands;
-                if (operands2 == null || operands2.Count < 2) goto callBase;
-
-                UnaryExpression ue = operands2[1] as UnaryExpression;
-                if (ue == null) goto callBase;
-
-                MemberBinding memberBinding = ue.Operand as MemberBinding;
-                if (memberBinding == null) goto callBase;
-
-                Method closure = memberBinding.BoundMember as Method;
-                if (closure == null || closure.Parameters == null || closure.Parameters.Count < 1) goto callBase;
-
-                closureParameter = closure.Parameters[0];
-
-            callBase:
-                if (closureParameter != null)
-                {
-                    this.stackOfMethods.Add(call); // imitate Stack.Push()
-                    this.stackOfBoundVariables.Add(closureParameter);
-                }
-
-                Expression result = base.VisitMethodCall(call);
-
-                if (closureParameter != null)
-                {
-                    this.stackOfMethods.RemoveAt(this.stackOfMethods.Count - 1); // imitate Stack.Pop()
-                    this.stackOfBoundVariables.RemoveAt(this.stackOfBoundVariables.Count - 1); // imitate Stack.Pop()
-                }
-
-                return result;
-            }
-
-            // All other method calls
-
-            return base.VisitMethodCall(call);
-        }
-
-        /// <summary>
-        /// If there is an anonymous delegate within a postcondition, then there
-        /// will be a call to a delegate constructor.
-        /// That call looks like "d..ctor(o,m)" where d is the type of the delegate.
-        /// There are two cases depending on whether the anonymous delegate captured
-        /// anything. In both cases, m is the method implementing the anonymous delegate.
-        /// 
-        /// (1) It does capture something. Then o is the instance of the closure class
-        /// implementing the delegate, and m is an instance method in the closure
-        /// class.
-        /// (2) It does *not* capture anything. Then o is the literal for null and
-        /// m is a static method that was added directly to the class.
-        /// 
-        /// This method will cause m to be visited to collect any old expressions
-        /// that occur in it. But those old expressions are not turned into locals,
-        /// but fields. The fields are created as either (1) members of the closure class,
-        /// or (2) members of the class in which the method containing the contract
-        /// containing the Old expression is defined.
-        /// The fields are initialized when (1) the closure class instance is created,
-        /// or (2) ... ??? (when!?)
-        /// So set up enough state so that when those old expressions are visited, the
-        /// correct ASTs can be built.
-        /// </summary>
-        /// <param name="cons">The AST representing the call to the constructor
-        /// of the delegate type.</param>
-        /// <returns>Whatever the base visitor returns</returns>
-        public override Expression VisitConstruct(Construct cons)
-        {
-            if (cons.Type is DelegateNode)
-            {
-                UnaryExpression ue = cons.Operands[1] as UnaryExpression;
-                if (ue == null) goto JustVisit;
-
-                MemberBinding mb = ue.Operand as MemberBinding;
-                if (mb == null) goto JustVisit;
-
-                Method m = mb.BoundMember as Method;
-
-                Contract.Assume(m != null);
-
-                if (m.IsStatic)
-                {
-                    // then there is no closure class, m is just a static method the compiler
-                    // added to the class itself
-                    goto JustVisit;
-                }
-                if (HelperMethods.IsCompilerGenerated(m))
-                {
-                    Local l = cons.Operands[0] as Local;
-                    if (l == null) goto JustVisit; // but then what is it??
-
-                    TypeNode savedClosureClass = this.currentClosureClass;
-
-                    this.currentClosureClass = l.Type;
-                    if (savedClosureClass == null)
-                    {
-                        // then this is the top-level closure class
-                        // have to treat it special: it doesn't contain a field that points 
-                        // to the top-level closure class. The field introduced to hold the value of the
-                        // old expression will be declared in this class.
-                        this.topLevelClosureClass = this.currentClosureClass;
-                    }
-                    else
-                    {
-                        // Find the field in this.closureClass that the C# compiler generated
-                        // to point to the top-level closure
-                        foreach (Member mem in this.currentClosureClass.Members)
-                        {
-                            Field f = mem as Field;
-                            if (f == null) continue;
-
-                            if (f.Type == this.topLevelClosureClass)
-                            {
-                                this.PointerToTopLevelClosureClass = f;
-                                break;
-                            }
-                        }
-                    }
-
-                    this.VisitBlock(m.Body);
-
-                    if (savedClosureClass == null)
-                    {
-                        this.topLevelClosureClass = null;
-                    }
-
-                    this.currentClosureClass = savedClosureClass;
-                    this.PointerToTopLevelClosureClass = null;
-                }
-            }
-
-        JustVisit:
-            return base.VisitConstruct(cons);
-        }
-    }
-
-    internal sealed class SubstituteClosureClassWithinOldExpressions : StandardVisitor
-    {
-        private readonly Dictionary<TypeNode, Local> closureLocals;
-
-        public SubstituteClosureClassWithinOldExpressions(Dictionary<TypeNode, Local> closureLocals)
-        {
-            this.closureLocals = closureLocals;
-        }
-
-        public override Expression VisitMemberBinding(MemberBinding memberBinding)
-        {
-            Local closureLocal;
-            if (memberBinding.TargetObject != null &&
-                this.closureLocals.TryGetValue(memberBinding.TargetObject.Type, out closureLocal))
-            {
-                return new MemberBinding(closureLocal, memberBinding.BoundMember);
-            }
-            
-            return base.VisitMemberBinding(memberBinding);
-        }
-    }
-
-    internal sealed class SubstituteParameters : StandardVisitor
-    {
-        private readonly TrivialHashtable map;
-        private readonly List<Parameter> parameters;
-
-        public SubstituteParameters(TrivialHashtable map, List<Parameter> parameters)
-        {
-            this.map = map;
-            this.parameters = parameters;
-        }
-
-        public override Expression VisitMemberBinding(MemberBinding memberBinding)
-        {
-            if (memberBinding.TargetObject != null &&
-                (memberBinding.TargetObject.NodeType == NodeType.This
-                 || memberBinding.TargetObject.NodeType == NodeType.Local))
-            {
-                // search in list of parameters to see if any have the same name as the bound member
-                foreach (Parameter p in this.parameters)
-                {
-                    if (p.Name.UniqueIdKey == memberBinding.BoundMember.Name.UniqueIdKey)
-                    {
-                        return (Expression) this.map[p.UniqueKey];
-                    }
-                }
-
-                return base.VisitMemberBinding(memberBinding);
-            }
-            
-            return base.VisitMemberBinding(memberBinding);
-        }
-
-        public override Expression VisitParameter(Parameter parameter)
-        {
-            if (parameter == null) return null;
-
-            object result = map[parameter.UniqueKey];
-            if (result != null) return (Expression) result;
-
-            return base.VisitParameter(parameter);
-        }
-    }
-
-    /// <summary>
-    /// Visitor is used for invariants only!
-    /// </summary>
-    internal sealed class RewriteInvariant : StandardVisitor
-    {
-        private readonly RuntimeContractMethods rcm;
-        private Literal sourceTextOfInvariant = null;
-
-        // Requires:
-        //  replacementMethod.Keys == replacementExceptions.Keys
-        //  
-        public RewriteInvariant(RuntimeContractMethods rcm)
-        {
-            this.rcm = rcm;
-        }
-
-        // Requires:
-        //  statement.Expression is MethodCall
-        //  statement.Expression.Callee is MemberBinding
-        //  statement.Expression.Callee.BoundMember is Method
-        //  statement.Expression.Callee.BoundMember == "Requires" or "Ensures"
-        //
-        //  inline  <==> replacementMethod == null
-        //  replacementMethod != null
-        //           ==> replacementMethod.ReturnType == methodToReplace.ReturnType
-        //               && replacementMethod.Parameters.Count == 1
-        //               && methodToReplace.Parameters.Count == 1
-        //               && replacementMethod.Parameters[0].Type == methodToReplace.Parameters[0].Type
-        //
-        private static Statement RewriteContractCall(
-            ExpressionStatement statement,
-            Method /*!*/ methodToReplace,
-            Method /*?*/ replacementMethod,
-            Literal /*?*/ sourceTextToUseAsSecondArg)
-        {
-            Contract.Requires(statement != null);
-
-            MethodCall call = statement.Expression as MethodCall;
-
-            if (call == null || call.Callee == null)
-            {
-                return statement;
-            }
-
-            MemberBinding mb = call.Callee as MemberBinding;
-            if (mb == null)
-            {
-                return statement;
-            }
-
-            Method m = mb.BoundMember as Method;
-            if (m == null)
-            {
-                return statement;
-            }
-
-            if (m != methodToReplace)
-            {
-                return statement;
-            }
-
-            mb.BoundMember = replacementMethod;
-
-            if (call.Operands.Count == 3)
-            {
-                // then the invariant was found in a reference assembly
-                // it already has all of its arguments
-                return statement;
-            }
-
-            if (call.Operands.Count == 1)
-            {
-                call.Operands.Add(Literal.Null);
-            }
-
-            Literal extraArg = sourceTextToUseAsSecondArg;
-            if (extraArg == null)
-            {
-                extraArg = Literal.Null;
-                //extraArg = new Literal("No other information available", SystemTypes.String);
-            }
-
-            call.Operands.Add(extraArg);
-
-            return statement;
-        }
-
-        public override Statement VisitExpressionStatement(ExpressionStatement statement)
-        {
-            Method contractMethod = Rewriter.ExtractCallFromStatement(statement);
-
-            if (contractMethod == null)
-            {
-                return base.VisitExpressionStatement(statement);
-            }
-
-            if (this.rcm.ContractNodes.IsInvariantMethod(contractMethod))
-            {
-                return RewriteContractCall(statement, contractMethod, this.rcm.InvariantMethod, this.sourceTextOfInvariant);
-            }
-
-            return base.VisitExpressionStatement(statement);
-        }
-
-        public override Invariant VisitInvariant(Invariant @invariant)
-        {
-            if (@invariant == null) return null;
-
-            SourceContext sctx = @invariant.SourceContext;
-            if (invariant.SourceConditionText != null)
-            {
-                this.sourceTextOfInvariant = invariant.SourceConditionText;
-            }
-            else if (sctx.IsValid && sctx.Document.Text != null && sctx.Document.Text.Source != null)
-            {
-                this.sourceTextOfInvariant = new Literal(sctx.Document.Text.Source, SystemTypes.String);
-            }
-            else
-            {
-                this.sourceTextOfInvariant = Literal.Null;
-                //this.sourceTextOfInvariant = new Literal("No other information available", SystemTypes.String);
-            }
-
-            return base.VisitInvariant(@invariant);
-        }
-    }
-
-    internal sealed class RemoveShortBranches : Inspector
-    {
-        public override void VisitBranch(Branch branch)
-        {
-            branch.ShortOffset = false;
-        }
-    }
-
-    internal sealed class WrapParametersInOldExpressions : StandardVisitor
-    {
-        public override Expression VisitOldExpression(OldExpression oldExpression)
-        {
-            // no point descending down into any old expressions
-            return oldExpression;
-        }
-
-        public override Expression VisitParameter(Parameter parameter)
-        {
-            if (parameter == null) return null;
-
-            var oe = new OldExpression(parameter);
-            oe.Type = parameter.Type;
-
-            return oe;
-        }
-    }
-
-    /// <summary>
-    /// Encapsulates the runtime methods to use when emitting checks.
-    /// Generates necessary code on demand.
-    /// </summary>
-    public sealed class RuntimeContractMethods
-    {
-        public readonly bool ThrowOnFailure;
-        public readonly int RewriteLevel;
-        public readonly bool PublicSurfaceOnly;
-        public readonly bool CallSiteRequires;
-        private readonly int regularRecursionGuard;
-        public readonly bool HideFromDebugger;
-
-        public bool AssertOnFailure
-        {
-            get { return !ThrowOnFailure; }
-        }
-
-        /// <summary>
-        /// Controls how validations are emitted and how inheritance works for all requires that throw exceptions.
-        /// </summary>
-        public readonly bool UseExplicitValidation;
-
-        public RuntimeContractMethods(TypeNode userContractType, ContractNodes contractNodes,
-            AssemblyNode targetAssembly,
-            bool throwOnFailure, int rewriteLevel, bool publicSurfaceOnly, bool callSiteRequires,
-            int recursionGuard, bool hideFromDebugger,
-            bool userExplicitValidation)
-        {
-            this.contractNodes = contractNodes;
-            this.targetAssembly = targetAssembly;
-            this.ThrowOnFailure = throwOnFailure;
-            this.RewriteLevel = rewriteLevel;
-            this.PublicSurfaceOnly = publicSurfaceOnly;
-            this.CallSiteRequires = callSiteRequires;
-            this.regularRecursionGuard = recursionGuard;
-            this.HideFromDebugger = hideFromDebugger;
-            this.UseExplicitValidation = userExplicitValidation;
-
-            // extract methods from user methods
-
-            // Get the user-specified rewriter methods (optional) REVIEW!! Needs a lot of error handling
-
-            if (userContractType != null)
-            {
-                Method method = null;
-                MethodList reqMethods = userContractType.GetMethods(Identifier.For("Requires"), SystemTypes.Boolean,
-                    SystemTypes.String, SystemTypes.String);
-
-                for (int i = 0; i < reqMethods.Count; i++)
-                {
-                    method = reqMethods[i];
-                    if (method != null)
-                    {
-                        if (method.TemplateParameters == null || method.TemplateParameters.Count != 1)
-                        {
-                            /*if (method != null) */
-                            this.requiresMethod = method;
-                        }
-                        else
-                        {
-                            this.requiresWithExceptionMethod = method;
-                        }
-                    }
-                }
-
-                method = userContractType.GetMethod(Identifier.For("Ensures"), SystemTypes.Boolean, SystemTypes.String, SystemTypes.String);
-                if (method != null) this.ensuresMethod = method;
-
-                method = userContractType.GetMethod(Identifier.For("EnsuresOnThrow"), SystemTypes.Boolean, SystemTypes.String, SystemTypes.String, SystemTypes.Exception);
-                if (method != null) this.ensuresOnThrowMethod = method;
-
-                method = userContractType.GetMethod(Identifier.For("Invariant"), SystemTypes.Boolean, SystemTypes.String, SystemTypes.String);
-                if (method != null) this.invariantMethod = method;
-
-                method = userContractType.GetMethod(Identifier.For("Assert"), SystemTypes.Boolean, SystemTypes.String, SystemTypes.String);
-                if (method != null) this.assertMethod = method;
-
-                method = userContractType.GetMethod(Identifier.For("Assume"), SystemTypes.Boolean, SystemTypes.String, SystemTypes.String);
-                if (method != null) this.assumeMethod = method;
-
-                // Need to make sure that the type ContractFailureKind is the one used in the user-supplied methods, which is not necessarily
-                // the one that is defined in the assembly that defines the contract class. For instance, extracting/rewriting from a 4.0 assembly
-                // but where the user-supplied assembly is pre-4.0.
-                var mems = userContractType.GetMembersNamed(ContractNodes.ReportFailureName);
-                TypeNode contractFailureKind = contractNodes.ContractFailureKind;
-
-                //if (mems != null) 
-                {
-                    foreach (var mem in mems)
-                    {
-                        method = mem as Method;
-                        if (method == null) continue;
-
-                        if (method.Parameters.Count != 4) continue;
-
-                        if (method.Parameters[0].Type.Name != contractNodes.ContractFailureKind.Name) continue;
-
-                        if (method.Parameters[1].Type != SystemTypes.String) continue;
-
-                        if (method.Parameters[2].Type != SystemTypes.String) continue;
-
-                        if (method.Parameters[3].Type != SystemTypes.Exception) continue;
-
-                        this.failureMethod = method;
-                        contractFailureKind = method.Parameters[0].Type;
-
-                        break;
-                    }
-                }
-
-                if (this.failureMethod == null)
-                {
-                    mems = userContractType.GetMembersNamed(ContractNodes.RaiseContractFailedEventName);
-                    // if (mems != null) 
-                    {
-                        foreach (var mem in mems)
-                        {
-                            method = mem as Method;
-                            if (method == null) continue;
-
-                            if (method.Parameters.Count != 4) continue;
-
-                            if (method.Parameters[0].Type.Name.UniqueIdKey !=
-                                contractNodes.ContractFailureKind.Name.UniqueIdKey) continue;
-
-                            if (method.Parameters[1].Type != SystemTypes.String) continue;
-
-                            if (method.Parameters[2].Type != SystemTypes.String) continue;
-
-                            if (method.Parameters[3].Type != SystemTypes.Exception) continue;
-
-                            this.raiseFailureEventMethod = method;
-                            contractFailureKind = method.Parameters[0].Type;
-
-                            break;
-                        }
-                    }
-                }
-                else
-                {
-                    method = userContractType.GetMethod(ContractNodes.RaiseContractFailedEventName, contractFailureKind,
-                        SystemTypes.String, SystemTypes.String, SystemTypes.Exception);
-
-                    if (method != null) this.raiseFailureEventMethod = method;
-                }
-
-                if (this.raiseFailureEventMethod != null)
-                {
-                    // either take all both RaiseContractFailedEvent and TriggerFailure or neither
-                    method = userContractType.GetMethod(ContractNodes.TriggerFailureName, contractFailureKind,
-                        SystemTypes.String, SystemTypes.String, SystemTypes.String, SystemTypes.Exception);
-
-                    if (method != null) this.triggerFailureMethod = method;
-                }
-            }
-        }
-
-        public int RecursionGuardCountFor(Method method)
-        {
-            Contract.Requires(method != null);
-
-            if (method.IsPropertyGetter) return Math.Min(1, regularRecursionGuard);
-            return regularRecursionGuard;
-        }
-
-        // Runtime contract methods
-
-        private Method /*?*/ requiresMethod = null;
-        private Method /*?*/ requiresWithExceptionMethod = null;
-        private Method /*?*/ ensuresMethod = null;
-        private Method /*?*/ ensuresOnThrowMethod = null;
-        private Method /*?*/ invariantMethod = null;
-        private Method /*?*/ assertMethod = null;
-        private Method /*?*/ assumeMethod = null;
-        private Method /*?*/ failureMethod = null;
-
-        public Method RequiresMethod
-        {
-            get
-            {
-                if (requiresMethod == null)
-                {
-                    // generate it
-                    this.requiresMethod = MakeMethod("Requires", ContractFailureKind.Precondition);
-                }
-
-                return requiresMethod;
-            }
-        }
-
-        public Method RequiresWithExceptionMethod
-        {
-            get
-            {
-                if (this.requiresWithExceptionMethod == null)
-                {
-                    this.requiresWithExceptionMethod = MakeRequiresWithExceptionMethod("Requires");
-                }
-
-                return this.requiresWithExceptionMethod;
-            }
-        }
-
-        public Method EnsuresMethod
-        {
-            get
-            {
-                if (ensuresMethod == null)
-                {
-                    // generate it
-                    this.ensuresMethod = MakeMethod("Ensures", ContractFailureKind.Postcondition);
-                }
-
-                return ensuresMethod;
-            }
-        }
-
-        public Method EnsuresOnThrowMethod
-        {
-            get
-            {
-                if (this.ensuresOnThrowMethod == null)
-                {
-                    // generate it
-                    this.ensuresOnThrowMethod = MakeMethod("EnsuresOnThrow", ContractFailureKind.PostconditionOnException);
-                }
-
-                return this.ensuresOnThrowMethod;
-            }
-        }
-
-        public Method InvariantMethod
-        {
-            get
-            {
-                if (this.invariantMethod == null)
-                {
-                    // generate it
-                    this.invariantMethod = MakeMethod("Invariant", ContractFailureKind.Invariant);
-                }
-
-                return this.invariantMethod;
-            }
-        }
-
-        public Method AssertMethod
-        {
-            get
-            {
-                if (this.assertMethod == null)
-                {
-                    // generate it
-                    this.assertMethod = MakeMethod("Assert", ContractFailureKind.Assert);
-                }
-
-                return this.assertMethod;
-            }
-        }
-
-        public Method AssumeMethod
-        {
-            get
-            {
-                if (this.assumeMethod == null)
-                {
-                    // generate it
-                    this.assumeMethod = MakeMethod("Assume", ContractFailureKind.Assume);
-                }
-
-                return this.assumeMethod;
-            }
-        }
-
-        // FailureKind literals
-
-        internal Literal PreconditionKind
-        {
-            get { return contractNodes.ContractFailureKind.GetField(Identifier.For("Precondition")).DefaultValue; }
-        }
-
-        // other private fields
-
-        private Class /*?*/ runtimeContractType;
-
-        private readonly ContractNodes contractNodes;
-        private readonly AssemblyNode targetAssembly;
-
-        public ContractNodes ContractNodes
-        {
-            get { return this.contractNodes; }
-        }
-
-        private Method FailureMethod
-        {
-            get
-            {
-                if (this.failureMethod == null)
-                {
-                    // Generate it.
-                    Method m = MakeFailureMethod();
-
-                    this.failureMethod = m;
-                }
-
-                return this.failureMethod;
-            }
-        }
-
-
-        private Method raiseFailureEventMethod;
-
-        internal Method RaiseFailureEventMethod
-        {
-            get
-            {
-                if (this.raiseFailureEventMethod == null)
-                {
-                    this.raiseFailureEventMethod = this.contractNodes.RaiseFailedEventMethod;
-                }
-
-                if (this.raiseFailureEventMethod == null)
-                {
-                    // Generate it
-                    Method m = MakeRaiseFailureEventMethod();
-
-                    this.raiseFailureEventMethod = m;
-                }
-
-                return this.raiseFailureEventMethod;
-            }
-        }
-
-        private Method triggerFailureMethod;
-
-        private Method TriggerFailureMethod
-        {
-            get
-            {
-                if (this.triggerFailureMethod == null && !ThrowOnFailure)
-                {
-                    this.triggerFailureMethod = this.contractNodes.TriggerFailureMethod;
-                }
-
-                if (this.triggerFailureMethod == null)
-                {
-                    // Generate it
-                    Method m = MakeTriggerFailureMethod(ThrowOnFailure);
-
-                    this.triggerFailureMethod = m;
-                }
-
-                return this.triggerFailureMethod;
-            }
-        }
-
-        private Class contractExceptionType;
-
-        private Class ContractExceptionType
-        {
-            get
-            {
-                if (this.contractExceptionType == null)
-                {
-                    if (this.contractNodes.ContractClass.DeclaringModule == this.targetAssembly)
-                    {
-                        // If we're rewriting an assembly that defines the contract class itself,
-                        // see if it already defines a contract exception that can be used
-                        this.contractExceptionType =
-                            this.targetAssembly.GetType(
-                                ContractNodes.ContractNamespace,
-                                Identifier.For("ContractException")) as Class;
-                    }
-
-                    // If that fails for any reason, then go ahead and make our own contract exception type
-                    if (this.contractExceptionType == null)
-                    {
-                        this.contractExceptionType = MakeContractException();
-                    }
-                }
-
-                return this.contractExceptionType;
-            }
-        }
-
-        private Class MakeContractException()
-        {
-            Class contractExceptionType;
-
-            // If we're rewriting an assembly for v4 or above and it *isn't* Silverlight (so serialization support is needed), then use new embedded dll as the type
-
-            if (4 <= TargetPlatform.MajorVersion)
-            {
-                var iSafeSerializationData =
-                    SystemTypes.SystemAssembly.GetType(Identifier.For("System.Runtime.Serialization"),
-                        Identifier.For("ISafeSerializationData")) as Interface;
-
-                if (iSafeSerializationData != null)
-                {
-                    // Just much easier to write the C# and have the compiler generate everything than to try and create it all manually
-                    System.Reflection.Assembly embeddedAssembly;
-
-                    Stream embeddedAssemblyStream;
-                    embeddedAssembly = System.Reflection.Assembly.GetExecutingAssembly();
-                    embeddedAssemblyStream =
-                        embeddedAssembly.GetManifestResourceStream("Microsoft.Contracts.Foxtrot.InternalException.dll");
-
-                    byte[] data = new byte[0];
-                    using (var br = new BinaryReader(embeddedAssemblyStream))
-                    {
-                        var len = embeddedAssemblyStream.Length;
-                        if (len < Int32.MaxValue)
-                            data = br.ReadBytes((int) len);
-
-                        AssemblyNode assemblyNode = AssemblyNode.GetAssembly(data, TargetPlatform.StaticAssemblyCache, true, false, true);
-                        contractExceptionType = assemblyNode.GetType(Identifier.For(""), Identifier.For("ContractException")) as Class;
-                    }
-
-                    if (contractExceptionType == null)
-                        throw new RewriteException("Tried to create the ContractException type from the embedded dll, but failed");
-
-                    var d = new Duplicator(this.targetAssembly, this.RuntimeContractType);
-                    d.FindTypesToBeDuplicated(new TypeNodeList(contractExceptionType));
-
-                    var ct = d.Visit(contractExceptionType);
-
-                    contractExceptionType = (Class) ct;
-                    contractExceptionType.Flags |= TypeFlags.NestedPrivate;
-
-                    this.RuntimeContractType.Members.Add(contractExceptionType);
-                    return contractExceptionType;
-                }
-            }
-
-            contractExceptionType = new Class(this.targetAssembly, this.RuntimeContractType, new AttributeList(),
-                TypeFlags.Class | TypeFlags.NestedPrivate | TypeFlags.Serializable, null,
-                Identifier.For("ContractException"), SystemTypes.Exception, null, null);
-
-            RewriteHelper.TryAddCompilerGeneratedAttribute(contractExceptionType);
-
-            var kindField = new Field(contractExceptionType, null, FieldFlags.Private, Identifier.For("_Kind"), contractNodes.ContractFailureKind, null);
-            var userField = new Field(contractExceptionType, null, FieldFlags.Private, Identifier.For("_UserMessage"), SystemTypes.String, null);
-            var condField = new Field(contractExceptionType, null, FieldFlags.Private, Identifier.For("_Condition"), SystemTypes.String, null);
-
-            contractExceptionType.Members.Add(kindField);
-            contractExceptionType.Members.Add(userField);
-            contractExceptionType.Members.Add(condField);
-
-            // Constructor for setting the fields
-
-            var parameters = new ParameterList();
-
-            var kindParam = new Parameter(Identifier.For("kind"), this.contractNodes.ContractFailureKind);
-            var failureParam = new Parameter(Identifier.For("failure"), SystemTypes.String);
-            var usermsgParam = new Parameter(Identifier.For("usermsg"), SystemTypes.String);
-            var conditionParam = new Parameter(Identifier.For("condition"), SystemTypes.String);
-            var innerParam = new Parameter(Identifier.For("inner"), SystemTypes.Exception);
-
-            parameters.Add(kindParam);
-            parameters.Add(failureParam);
-            parameters.Add(usermsgParam);
-            parameters.Add(conditionParam);
-            parameters.Add(innerParam);
-
-            var body = new Block(new StatementList());
-
-            var ctor = new InstanceInitializer(contractExceptionType, null, parameters, body);
-            ctor.Flags |= MethodFlags.Public | MethodFlags.HideBySig;
-            ctor.CallingConvention = CallingConventionFlags.HasThis;
-
-            body.Statements.Add(
-                new ExpressionStatement(
-                    new MethodCall(
-                        new MemberBinding(ctor.ThisParameter,
-                            contractExceptionType.BaseClass.GetConstructor(SystemTypes.String, SystemTypes.Exception)),
-                        new ExpressionList(failureParam, innerParam))));
-
-            body.Statements.Add(new AssignmentStatement(new MemberBinding(ctor.ThisParameter, kindField), kindParam));
-            body.Statements.Add(new AssignmentStatement(new MemberBinding(ctor.ThisParameter, userField), usermsgParam));
-            body.Statements.Add(new AssignmentStatement(new MemberBinding(ctor.ThisParameter, condField), conditionParam));
-
-            body.Statements.Add(new Return());
-
-            contractExceptionType.Members.Add(ctor);
-
-            if (SystemTypes.SerializationInfo != null && SystemTypes.SerializationInfo.BaseClass != null)
-            {
-                // Silverlight (e.g.) is a platform that doesn't support serialization. So check to make sure the type really exists.
-                // 
-                var baseCtor = SystemTypes.Exception.GetConstructor(SystemTypes.SerializationInfo,
-                    SystemTypes.StreamingContext);
-
-                if (baseCtor != null)
-                {
-                    // Deserialization Constructor
-
-                    parameters = new ParameterList();
-                    var info = new Parameter(Identifier.For("info"), SystemTypes.SerializationInfo);
-                    var context = new Parameter(Identifier.For("context"), SystemTypes.StreamingContext);
-
-                    parameters.Add(info);
-                    parameters.Add(context);
-
-                    body = new Block(new StatementList());
-                    ctor = new InstanceInitializer(contractExceptionType, null, parameters, body);
-
-                    ctor.Flags |= MethodFlags.Private | MethodFlags.HideBySig;
-                    ctor.CallingConvention = CallingConventionFlags.HasThis;
-
-                    // : base(info, context) 
-                    body.Statements.Add(
-                        new ExpressionStatement(new MethodCall(new MemberBinding(ctor.ThisParameter, baseCtor),
-                            new ExpressionList(info, context))));
-
-                    // _Kind = (ContractFailureKind)info.GetInt32("Kind");
-                    var getInt32 = SystemTypes.SerializationInfo.GetMethod(Identifier.For("GetInt32"), SystemTypes.String);
-
-                    body.Statements.Add(new AssignmentStatement(
-                        new MemberBinding(new This(), kindField),
-                        new MethodCall(new MemberBinding(info, getInt32),
-                            new ExpressionList(new Literal("Kind", SystemTypes.String)))
-                        ));
-
-                    // _UserMessage = info.GetString("UserMessage");
-                    var getString = SystemTypes.SerializationInfo.GetMethod(Identifier.For("GetString"), SystemTypes.String);
-                    body.Statements.Add(new AssignmentStatement(
-                        new MemberBinding(new This(), userField),
-                        new MethodCall(new MemberBinding(info, getString),
-                            new ExpressionList(new Literal("UserMessage", SystemTypes.String)))
-                        ));
-
-                    // _Condition = info.GetString("Condition");
-                    body.Statements.Add(new AssignmentStatement(
-                        new MemberBinding(new This(), condField),
-                        new MethodCall(new MemberBinding(info, getString),
-                            new ExpressionList(new Literal("Condition", SystemTypes.String)))
-                        ));
-
-                    body.Statements.Add(new Return());
-                    contractExceptionType.Members.Add(ctor);
-
-                    // GetObjectData
-
-                    var securityCriticalCtor = SystemTypes.SecurityCriticalAttribute.GetConstructor();
-                    var securityCriticalAttribute = new AttributeNode(new MemberBinding(null, securityCriticalCtor),
-                        null, System.AttributeTargets.Method);
-
-                    var attrs = new AttributeList(securityCriticalAttribute);
-                    parameters = new ParameterList();
-
-                    info = new Parameter(Identifier.For("info"), SystemTypes.SerializationInfo);
-                    context = new Parameter(Identifier.For("context"), SystemTypes.StreamingContext);
-
-                    parameters.Add(info);
-                    parameters.Add(context);
-
-                    body = new Block(new StatementList());
-                    var getObjectDataName = Identifier.For("GetObjectData");
-                    var getObjectData = new Method(contractExceptionType, attrs, getObjectDataName, parameters,
-                        SystemTypes.Void, body);
-
-                    getObjectData.CallingConvention = CallingConventionFlags.HasThis;
-
-                    // public override
-                    getObjectData.Flags = MethodFlags.Public | MethodFlags.Virtual;
-
-                    // base.GetObjectData(info, context);
-                    var baseGetObjectData = SystemTypes.Exception.GetMethod(getObjectDataName,
-                        SystemTypes.SerializationInfo, SystemTypes.StreamingContext);
-                    body.Statements.Add(new ExpressionStatement(
-                        new MethodCall(new MemberBinding(new This(), baseGetObjectData),
-                            new ExpressionList(info, context), NodeType.Call, SystemTypes.Void)
-                        ));
-
-                    // info.AddValue("Kind", _Kind);
-                    var addValueObject = SystemTypes.SerializationInfo.GetMethod(Identifier.For("AddValue"),
-                        SystemTypes.String, SystemTypes.Object);
-                    body.Statements.Add(new ExpressionStatement(
-                        new MethodCall(new MemberBinding(info, addValueObject),
-                            new ExpressionList(new Literal("Kind", SystemTypes.String),
-                                new BinaryExpression(new MemberBinding(new This(), kindField),
-                                    new Literal(contractNodes.ContractFailureKind), NodeType.Box)), NodeType.Call,
-                            SystemTypes.Void)
-                        ));
-
-                    // info.AddValue("UserMessage", _UserMessage);
-                    body.Statements.Add(new ExpressionStatement(
-                        new MethodCall(new MemberBinding(info, addValueObject),
-                            new ExpressionList(new Literal("UserMessage", SystemTypes.String),
-                                new MemberBinding(new This(), userField)), NodeType.Call, SystemTypes.Void)
-                        ));
-
-                    // info.AddValue("Condition", _Condition);
-                    body.Statements.Add(new ExpressionStatement(
-                        new MethodCall(new MemberBinding(info, addValueObject),
-                            new ExpressionList(new Literal("Condition", SystemTypes.String),
-                                new MemberBinding(new This(), condField)), NodeType.Call, SystemTypes.Void)
-                        ));
-
-                    body.Statements.Add(new Return());
-                    contractExceptionType.Members.Add(getObjectData);
-                }
-            }
-
-            this.RuntimeContractType.Members.Add(contractExceptionType);
-            return contractExceptionType;
-        }
-
-
-        private Method ContractExceptionCtor
-        {
-            get
-            {
-                TypeNode contractException = this.ContractExceptionType;
-
-                var result = contractException.GetConstructor(contractNodes.ContractFailureKind, SystemTypes.String,
-                    SystemTypes.String, SystemTypes.String, SystemTypes.Exception);
-
-                if (result == null)
-                {
-                    throw new ApplicationException(
-                        "Can't find constructor ContractException(ContractFailureKind, string, string, string, Exception).");
-                }
-
-                return result;
-            }
-        }
-
-        private TypeNode RuntimeContractType
-        {
-            get
-            {
-                Contract.Ensures(Contract.Result<TypeNode>() != null);
-
-                if (this.runtimeContractType == null)
-                {
-                    // generate
-                    this.runtimeContractType = new Class(this.targetAssembly,
-                        null, /* declaringType */
-                        new AttributeList(),
-                        TypeFlags.Abstract | TypeFlags.Sealed,
-                        ContractNodes.ContractNamespace,
-// for debugging            Identifier.For("__ContractsRuntime" + this.targetAssembly.UniqueKey),
-                        Identifier.For("__ContractsRuntime"),
-                        SystemTypes.Object,
-                        new InterfaceList(),
-                        new MemberList(0));
-
-                    RewriteHelper.TryAddCompilerGeneratedAttribute(runtimeContractType);
-                }
-
-                return runtimeContractType;
-            }
-        }
-
-        private Field inContractEvaluationField;
-
-        internal Field InContractEvaluationField
-        {
-            get
-            {
-                if (inContractEvaluationField == null)
-                {
-                    inContractEvaluationField = new Field(RuntimeContractType, null,
-                        FieldFlags.Assembly | FieldFlags.Static, Identifier.For("insideContractEvaluation"),
-                        SystemTypes.Int32, null);
-
-                    TypeNode threadStaticAttribute = SystemTypes.SystemAssembly.GetType(Identifier.For("System"),
-                        Identifier.For("ThreadStaticAttribute"));
-
-                    if (threadStaticAttribute != null)
-                    {
-                        inContractEvaluationField.Attributes.Add(
-                            new AttributeNode(new MemberBinding(null, threadStaticAttribute.GetConstructor()), null));
-                    }
-
-                    runtimeContractType.Members.Add(inContractEvaluationField);
-                }
-
-                return inContractEvaluationField;
-            }
-        }
-
-        /// <summary>
-        /// Constructs (and returns) a method that looks like this:
-        /// 
-        ///   [System.Diagnostics.DebuggerNonUserCodeAttribute]
-        ///   [System.Runtime.ConstrainedExecution.ReliabilityContractReliabilityContract(Consistency.WillNotCorruptState, Cer.MayFail)]
-        ///   static void name(bool condition, string message, string conditionText){
-        ///     if (!condition) {
-        ///       System.Diagnostics.Contracts.Contract.Failure(kind, message, conditionText, null);
-        ///     }
-        ///   }
-        ///
-        /// Or, if the ContractFailureKind is PostconditionOnException, then the generated method
-        /// gets an extra parameter which is an Exception and that parameter is passed to Contract.Failure instead of null
-        /// </summary>
-        private Method MakeMethod(string name, ContractFailureKind kind)
-        {
-            Parameter conditionParameter = new Parameter(Identifier.For("condition"), SystemTypes.Boolean);
-            Parameter messageParameter = new Parameter(Identifier.For("msg"), SystemTypes.String);
-            Parameter conditionTextParameter = new Parameter(Identifier.For("conditionTxt"), SystemTypes.String);
-            Parameter exceptionParameter = new Parameter(Identifier.For("originalException"), SystemTypes.Exception);
-
-            Block returnBlock = new Block(new StatementList(new Return()));
-
-            Block body = new Block(new StatementList());
-            Block b = new Block(new StatementList());
-
-            b.Statements.Add(new Branch(conditionParameter, returnBlock));
-            ExpressionList elist = new ExpressionList();
-
-            elist.Add(TranslateKindForBackwardCompatibility(kind));
-            elist.Add(messageParameter);
-            elist.Add(conditionTextParameter);
-
-            if (kind == ContractFailureKind.PostconditionOnException)
-            {
-                elist.Add(exceptionParameter);
-            }
-            else
-            {
-                elist.Add(Literal.Null);
-            }
-
-            b.Statements.Add(new ExpressionStatement(new MethodCall(new MemberBinding(null, this.FailureMethod), elist)));
-            b.Statements.Add(returnBlock);
-
-            body.Statements.Add(b);
-
-            ParameterList pl = new ParameterList(conditionParameter, messageParameter, conditionTextParameter);
-            if (kind == ContractFailureKind.PostconditionOnException)
-            {
-                pl.Add(exceptionParameter);
-            }
-
-            Method m = new Method(this.RuntimeContractType, null, Identifier.For(name), pl, SystemTypes.Void, body);
-            m.Flags = MethodFlags.Assembly | MethodFlags.Static;
-
-            m.Attributes = new AttributeList();
-            this.RuntimeContractType.Members.Add(m);
-
-            Member constructor = null;
-            AttributeNode attribute = null;
-
-            // Add [DebuggerNonUserCodeAttribute]
-
-            if (this.HideFromDebugger)
-            {
-                TypeNode debuggerNonUserCode = SystemTypes.SystemAssembly.GetType(Identifier.For("System.Diagnostics"),
-                    Identifier.For("DebuggerNonUserCodeAttribute"));
-
-                if (debuggerNonUserCode != null)
-                {
-                    constructor = debuggerNonUserCode.GetConstructor();
-                    attribute = new AttributeNode(new MemberBinding(null, constructor), null, AttributeTargets.Method);
-                    m.Attributes.Add(attribute);
-                }
-            }
-
-            TypeNode reliabilityContract =
-                SystemTypes.SystemAssembly.GetType(Identifier.For("System.Runtime.ConstrainedExecution"),
-                    Identifier.For("ReliabilityContractAttribute"));
-
-            TypeNode consistency =
-                SystemTypes.SystemAssembly.GetType(Identifier.For("System.Runtime.ConstrainedExecution"),
-                    Identifier.For("Consistency"));
-
-            TypeNode cer = SystemTypes.SystemAssembly.GetType(Identifier.For("System.Runtime.ConstrainedExecution"),
-                Identifier.For("Cer"));
-
-            if (reliabilityContract != null && consistency != null && cer != null)
-            {
-                constructor = reliabilityContract.GetConstructor(consistency, cer);
-                if (constructor != null)
-                {
-                    attribute = new AttributeNode(
-                        new MemberBinding(null, constructor),
-                        new ExpressionList(
-                            new Literal(System.Runtime.ConstrainedExecution.Consistency.WillNotCorruptState, consistency),
-                            new Literal(System.Runtime.ConstrainedExecution.Cer.MayFail, cer)),
-                        AttributeTargets.Method
-                        );
-
-                    m.Attributes.Add(attribute);
-                }
-            }
-
-            return m;
-        }
-
-        private Literal TranslateKindForBackwardCompatibility(ContractFailureKind kind)
-        {
-            switch (kind)
-            {
-                case ContractFailureKind.Assert:
-                {
-                    var field = this.contractNodes.ContractFailureKind.GetField(Identifier.For("Assert"));
-                    Contract.Assume(field != null);
-                    return field.DefaultValue;
-                }
-
-                case ContractFailureKind.Assume:
-                {
-                    var field = this.contractNodes.ContractFailureKind.GetField(Identifier.For("Assume"));
-                    Contract.Assume(field != null);
-                    return field.DefaultValue;
-                }
-
-                case ContractFailureKind.Invariant:
-                {
-                    var field = this.contractNodes.ContractFailureKind.GetField(Identifier.For("Invariant"));
-                    Contract.Assume(field != null);
-                    return field.DefaultValue;
-                }
-
-                case ContractFailureKind.Postcondition:
-                {
-                    var field = this.contractNodes.ContractFailureKind.GetField(Identifier.For("Postcondition"));
-                    Contract.Assume(field != null);
-                    return field.DefaultValue;
-                }
-
-                case ContractFailureKind.PostconditionOnException:
-                {
-                    var pe = this.contractNodes.ContractFailureKind.GetField(Identifier.For("PostconditionOnException"));
-                    // hack
-                    if (pe == null)
-                    {
-                        // Old CLR 4.0 beta
-                        var field = this.contractNodes.ContractFailureKind.GetField(Identifier.For("Postcondition"));
-                        Contract.Assume(field != null);
-                        return field.DefaultValue;
-                    }
-
-                    return pe.DefaultValue;
-                }
-
-                case ContractFailureKind.Precondition:
-                {
-                    var field = this.contractNodes.ContractFailureKind.GetField(Identifier.For("Precondition"));
-                    Contract.Assume(field != null);
-                    return field.DefaultValue;
-                }
-
-                default:
-                {
-                    throw new ApplicationException(String.Format("Unexpected failure kind {0}", kind.ToString()));
-                }
-            }
-        }
-
-        [ContractVerification(true)]
-        private Method MakeRequiresWithExceptionMethod(string name)
-        {
-            Parameter conditionParameter = new Parameter(Identifier.For("condition"), SystemTypes.Boolean);
-            Parameter messageParameter = new Parameter(Identifier.For("msg"), SystemTypes.String);
-            Parameter conditionTextParameter = new Parameter(Identifier.For("conditionTxt"), SystemTypes.String);
-
-            MethodClassParameter typeArg = new MethodClassParameter();
-            ParameterList pl = new ParameterList(conditionParameter, messageParameter, conditionTextParameter);
-
-            Block body = new Block(new StatementList());
-            
-            Method m = new Method(this.RuntimeContractType, null, Identifier.For(name), pl, SystemTypes.Void, body);
-            
-            m.Flags = MethodFlags.Assembly | MethodFlags.Static;
-            m.Attributes = new AttributeList();
-            m.TemplateParameters = new TypeNodeList(typeArg);
-            m.IsGeneric = true;
-
-            this.RuntimeContractType.Members.Add(m);
-
-            typeArg.Name = Identifier.For("TException");
-            typeArg.DeclaringMember = m;
-            typeArg.BaseClass = SystemTypes.Exception;
-            typeArg.ParameterListIndex = 0;
-            typeArg.DeclaringModule = this.RuntimeContractType.DeclaringModule;
-
-            Block returnBlock = new Block(new StatementList(new Return()));
-
-            Block b = new Block(new StatementList());
-            b.Statements.Add(new Branch(conditionParameter, returnBlock));
-
-            //
-            // Generate the following
-            //
-
-            // // message == null means: yes we handled it. Otherwise it is the localized failure message
-            // var message = RaiseFailureEvent(ContractFailureKind.Precondition, userMessage, conditionString, null);
-            // #if assertOnFailure
-            // if (message != null) {
-            //   Assert(false, message);
-            // }
-            // #endif
-
-            var messageLocal = new Local(Identifier.For("msg"), SystemTypes.String);
-
-            ExpressionList elist = new ExpressionList();
-
-            elist.Add(this.PreconditionKind);
-            elist.Add(messageParameter);
-            elist.Add(conditionTextParameter);
-            elist.Add(Literal.Null);
-
-            b.Statements.Add(new AssignmentStatement(messageLocal,
-                new MethodCall(new MemberBinding(null, this.RaiseFailureEventMethod), elist)));
-
-            if (this.AssertOnFailure)
-            {
-                var assertMethod = GetSystemDiagnosticsAssertMethod();
-                if (assertMethod != null)
-                {
-                    var skipAssert = new Block();
-
-                    b.Statements.Add(new Branch(new UnaryExpression(messageLocal, NodeType.LogicalNot), skipAssert));
-
-                    // emit assert call
-                    ExpressionList assertelist = new ExpressionList();
-                    assertelist.Add(Literal.False);
-                    assertelist.Add(messageLocal);
-
-                    b.Statements.Add(
-                        new ExpressionStatement(new MethodCall(new MemberBinding(null, assertMethod), assertelist)));
-
-                    b.Statements.Add(skipAssert);
-                }
-            }
-
-            // // construct exception
-            // Exception obj = null;
-            // ConstructorInfo ci = typeof(TException).GetConstructor(new[] { typeof(string), typeof(string) });
-            // if (ci != null)
-            // {
-            //   if (reflection.firstArgName == "paramName") {
-            //     exceptionObject = ci.Invoke(new[] { userMessage, message }) as Exception;
-            //   }
-            //   else {
-            //     exceptionObject = ci.Invoke(new[] { message, userMessage }) as Exception;
-            //   }
-            // }
-            // else {
-            //   ci = typeof(TException).GetConstructor(new[] { typeof(string) });
-            //   if (ci != null)
-            //   {
-            //     exceptionObject = ci.Invoke(new[] { message }) as Exception;
-            //   }
-            // }
-
-            var exceptionLocal = new Local(Identifier.For("obj"), SystemTypes.Exception);
-            b.Statements.Add(new AssignmentStatement(exceptionLocal, Literal.Null));
-
-            var constructorInfoType = TypeNode.GetTypeNode(typeof (System.Reflection.ConstructorInfo));
-
-            Contract.Assume(constructorInfoType != null);
-
-            var methodBaseType = TypeNode.GetTypeNode(typeof (System.Reflection.MethodBase));
-
-            Contract.Assume(methodBaseType != null);
-
-            var constructorLocal = new Local(Identifier.For("ci"), constructorInfoType);
-
-            Contract.Assume(SystemTypes.Type != null);
-
-            var getConstructorMethod = SystemTypes.Type.GetMethod(Identifier.For("GetConstructor"),
-                SystemTypes.Type.GetArrayType(1));
-
-            var typeofExceptionArg = GetTypeFromHandleExpression(typeArg);
-            var typeofString = GetTypeFromHandleExpression(SystemTypes.String);
-            var typeArrayLocal = new Local(Identifier.For("typeArray"), SystemTypes.Type.GetArrayType(1));
-            var typeArray2 = new ConstructArray();
-
-            typeArray2.ElementType = SystemTypes.Type;
-            typeArray2.Rank = 1;
-            typeArray2.Operands = new ExpressionList(Literal.Int32Two);
-            var typeArrayInit2 = new Block(new StatementList());
-
-            typeArrayInit2.Statements.Add(new AssignmentStatement(typeArrayLocal, typeArray2));
-            typeArrayInit2.Statements.Add(
-                new AssignmentStatement(
-                    new Indexer(typeArrayLocal, new ExpressionList(Literal.Int32Zero), SystemTypes.Type), typeofString));
-
-            typeArrayInit2.Statements.Add(
-                new AssignmentStatement(
-                    new Indexer(typeArrayLocal, new ExpressionList(Literal.Int32One), SystemTypes.Type), typeofString));
-
-            typeArrayInit2.Statements.Add(new ExpressionStatement(typeArrayLocal));
-            var typeArrayExpression2 = new BlockExpression(typeArrayInit2);
-            b.Statements.Add(new AssignmentStatement(constructorLocal,
-                new MethodCall(new MemberBinding(typeofExceptionArg, getConstructorMethod),
-                    new ExpressionList(typeArrayExpression2))));
-
-            var elseBlock = new Block();
-            b.Statements.Add(new Branch(new UnaryExpression(constructorLocal, NodeType.LogicalNot), elseBlock));
-            var endifBlock2 = new Block();
-
-            var invokeMethod = constructorInfoType.GetMethod(Identifier.For("Invoke"),
-                SystemTypes.Object.GetArrayType(1));
-
-            var argArray2 = new ConstructArray();
-            argArray2.ElementType = SystemTypes.Object;
-            argArray2.Rank = 1;
-            argArray2.Operands = new ExpressionList(Literal.Int32Two);
-            var argArrayLocal = new Local(Identifier.For("argArray"), SystemTypes.Object.GetArrayType(1));
-
-            var parameterInfoType = TypeNode.GetTypeNode(typeof (System.Reflection.ParameterInfo));
-
-            Contract.Assume(parameterInfoType != null);
-
-            var parametersMethod = methodBaseType.GetMethod(Identifier.For("GetParameters"));
-            var get_NameMethod = parameterInfoType.GetMethod(Identifier.For("get_Name"));
-            var string_op_EqualityMethod = SystemTypes.String.GetMethod(Identifier.For("op_Equality"),
-                SystemTypes.String, SystemTypes.String);
-
-            var elseArgMsgBlock = new Block();
-            var endIfArgMsgBlock = new Block();
-
-            b.Statements.Add(
-                new Branch(
-                    new UnaryExpression(
-                        new MethodCall(new MemberBinding(null, string_op_EqualityMethod),
-                            new ExpressionList(
-                                new MethodCall(
-                                    new MemberBinding(
-                                        new Indexer(
-                                            new MethodCall(new MemberBinding(constructorLocal, parametersMethod),
-                                                new ExpressionList(), NodeType.Callvirt),
-                                            new ExpressionList(Literal.Int32Zero), parameterInfoType), get_NameMethod),
-                                    new ExpressionList(), NodeType.Callvirt),
-                                new Literal("paramName", SystemTypes.String))), NodeType.LogicalNot), elseArgMsgBlock));
-
-            var argArrayInit2 = new Block(new StatementList());
-            argArrayInit2.Statements.Add(new AssignmentStatement(argArrayLocal, argArray2));
-            argArrayInit2.Statements.Add(
-                new AssignmentStatement(
-                    new Indexer(argArrayLocal, new ExpressionList(Literal.Int32Zero), SystemTypes.Object),
-                    messageParameter));
-
-            argArrayInit2.Statements.Add(
-                new AssignmentStatement(
-                    new Indexer(argArrayLocal, new ExpressionList(Literal.Int32One), SystemTypes.Object), messageLocal));
-            argArrayInit2.Statements.Add(new ExpressionStatement(argArrayLocal));
-
-            var argArrayExpression2 = new BlockExpression(argArrayInit2);
-            b.Statements.Add(new AssignmentStatement(exceptionLocal,
-                new BinaryExpression(
-                    new MethodCall(new MemberBinding(constructorLocal, invokeMethod),
-                        new ExpressionList(argArrayExpression2)), new Literal(SystemTypes.Exception), NodeType.Isinst)));
-            b.Statements.Add(new Branch(null, endIfArgMsgBlock));
-
-            b.Statements.Add(elseArgMsgBlock);
-
-            var argArrayInit2_1 = new Block(new StatementList());
-            argArrayInit2_1.Statements.Add(new AssignmentStatement(argArrayLocal, argArray2));
-            argArrayInit2_1.Statements.Add(
-                new AssignmentStatement(
-                    new Indexer(argArrayLocal, new ExpressionList(Literal.Int32Zero), SystemTypes.Object), messageLocal));
-
-            argArrayInit2_1.Statements.Add(
-                new AssignmentStatement(
-                    new Indexer(argArrayLocal, new ExpressionList(Literal.Int32One), SystemTypes.Object),
-                    messageParameter));
-            argArrayInit2_1.Statements.Add(new ExpressionStatement(argArrayLocal));
-
-            var argArrayExpression2_1 = new BlockExpression(argArrayInit2_1);
-            b.Statements.Add(new AssignmentStatement(exceptionLocal,
-                new BinaryExpression(
-                    new MethodCall(new MemberBinding(constructorLocal, invokeMethod),
-                        new ExpressionList(argArrayExpression2_1)), new Literal(SystemTypes.Exception), NodeType.Isinst)));
-
-            b.Statements.Add(endIfArgMsgBlock);
-
-            b.Statements.Add(new Branch(null, endifBlock2));
-
-            b.Statements.Add(elseBlock);
-
-            var typeArray1 = new ConstructArray();
-            typeArray1.ElementType = SystemTypes.Type;
-            typeArray1.Rank = 1;
-            typeArray1.Operands = new ExpressionList(Literal.Int32One);
-
-            var typeArrayInit1 = new Block(new StatementList());
-            typeArrayInit1.Statements.Add(new AssignmentStatement(typeArrayLocal, typeArray1));
-            typeArrayInit1.Statements.Add(
-                new AssignmentStatement(
-                    new Indexer(typeArrayLocal, new ExpressionList(Literal.Int32Zero), SystemTypes.Type), typeofString));
-
-            typeArrayInit1.Statements.Add(new ExpressionStatement(typeArrayLocal));
-            var typeArrayExpression1 = new BlockExpression(typeArrayInit1);
-
-            b.Statements.Add(new AssignmentStatement(constructorLocal,
-                new MethodCall(new MemberBinding(typeofExceptionArg, getConstructorMethod),
-                    new ExpressionList(typeArrayExpression1))));
-
-            b.Statements.Add(new Branch(new UnaryExpression(constructorLocal, NodeType.LogicalNot), endifBlock2));
-
-            var argArray1 = new ConstructArray();
-            argArray1.ElementType = SystemTypes.Object;
-            argArray1.Rank = 1;
-            argArray1.Operands = new ExpressionList(Literal.Int32One);
-
-            var argArrayInit1 = new Block(new StatementList());
-            argArrayInit1.Statements.Add(new AssignmentStatement(argArrayLocal, argArray1));
-            argArrayInit1.Statements.Add(
-                new AssignmentStatement(
-                    new Indexer(argArrayLocal, new ExpressionList(Literal.Int32Zero), SystemTypes.Object), messageLocal));
-
-            argArrayInit1.Statements.Add(new ExpressionStatement(argArrayLocal));
-            var argArrayExpression1 = new BlockExpression(argArrayInit1);
-
-            b.Statements.Add(new AssignmentStatement(exceptionLocal,
-                new BinaryExpression(
-                    new MethodCall(new MemberBinding(constructorLocal, invokeMethod),
-                        new ExpressionList(argArrayExpression1)), new Literal(SystemTypes.Exception), NodeType.Isinst)));
-
-            b.Statements.Add(endifBlock2);
-
-            // // throw it
-            // if (exceptionObject == null)
-            // {
-            //   throw new ArgumentException(displayMessage, message);
-            // }
-            // else
-            // {
-            //   throw exceptionObject;
-            // }
-
-            var thenBlock3 = new Block();
-
-            b.Statements.Add(new Branch(exceptionLocal, thenBlock3));
-            b.Statements.Add(
-                new Throw(
-                    new Construct(
-                        new MemberBinding(null,
-                            SystemTypes.ArgumentException.GetConstructor(SystemTypes.String, SystemTypes.String)),
-                        new ExpressionList(messageLocal, messageParameter))));
-
-            b.Statements.Add(thenBlock3);
-            b.Statements.Add(new Throw(exceptionLocal));
-            b.Statements.Add(returnBlock);
-
-            Contract.Assume(body.Statements != null);
-
-            body.Statements.Add(b);
-
-            Member constructor = null;
-            AttributeNode attribute = null;
-
-            // Add [DebuggerNonUserCodeAttribute]
-
-            if (this.HideFromDebugger)
-            {
-                TypeNode debuggerNonUserCode = SystemTypes.SystemAssembly.GetType(Identifier.For("System.Diagnostics"),
-                    Identifier.For("DebuggerNonUserCodeAttribute"));
-
-                if (debuggerNonUserCode != null)
-                {
-                    constructor = debuggerNonUserCode.GetConstructor();
-                    attribute = new AttributeNode(new MemberBinding(null, constructor), null, AttributeTargets.Method);
-                    m.Attributes.Add(attribute);
-                }
-            }
-
-            TypeNode reliabilityContract =
-                SystemTypes.SystemAssembly.GetType(Identifier.For("System.Runtime.ConstrainedExecution"),
-                    Identifier.For("ReliabilityContractAttribute"));
-
-            TypeNode consistency =
-                SystemTypes.SystemAssembly.GetType(Identifier.For("System.Runtime.ConstrainedExecution"),
-                    Identifier.For("Consistency"));
-
-            TypeNode cer = SystemTypes.SystemAssembly.GetType(Identifier.For("System.Runtime.ConstrainedExecution"),
-                Identifier.For("Cer"));
-
-            if (reliabilityContract != null && consistency != null && cer != null)
-            {
-                constructor = reliabilityContract.GetConstructor(consistency, cer);
-                if (constructor != null)
-                {
-                    attribute = new AttributeNode(
-                        new MemberBinding(null, constructor),
-                        new ExpressionList(
-                            new Literal(System.Runtime.ConstrainedExecution.Consistency.WillNotCorruptState, consistency),
-                            new Literal(System.Runtime.ConstrainedExecution.Cer.MayFail, cer)),
-                        AttributeTargets.Method
-                        );
-
-                    m.Attributes.Add(attribute);
-                }
-            }
-
-            return m;
-        }
-
-        internal static Method GetSystemDiagnosticsAssertMethod()
-        {
-            var sysref = (AssemblyReference) TargetPlatform.AssemblyReferenceFor[Identifier.For("System").UniqueIdKey];
-            if (sysref == null) return null;
-
-            var sysassem = AssemblyNode.GetAssembly(sysref, TargetPlatform.StaticAssemblyCache, true, false, true);
-            if (sysassem == null) return null;
-
-            var debugType = sysassem.GetType(Identifier.For("System.Diagnostics"), Identifier.For("Debug"));
-            if (debugType == null) return null;
-
-            var assertMethod = debugType.GetMethod(Identifier.For("Assert"), SystemTypes.Boolean, SystemTypes.String);
-            if (assertMethod == null) return null;
-
-            return assertMethod;
-        }
-
-        private static Expression GetTypeFromHandleExpression(TypeNode typeArg)
-        {
-            return
-                new MethodCall(
-                    new MemberBinding(null,
-                        SystemTypes.Type.GetMethod(Identifier.For("GetTypeFromHandle"), SystemTypes.RuntimeTypeHandle)),
-                    new ExpressionList(new UnaryExpression(new Literal(typeArg), NodeType.Ldtoken)));
-        }
-
-        private Method MakeFailureMethod()
-        {
-            Parameter kindParameter = new Parameter(Identifier.For("kind"), contractNodes.ContractFailureKind);
-            Parameter messageParameter = new Parameter(Identifier.For("msg"), SystemTypes.String);
-            Parameter conditionTextParameter = new Parameter(Identifier.For("conditionTxt"), SystemTypes.String);
-            Parameter exceptionParameter = new Parameter(Identifier.For("inner"), SystemTypes.Exception);
-            ParameterList pl = new ParameterList(kindParameter, messageParameter, conditionTextParameter, exceptionParameter);
-
-            Block body = new Block(new StatementList());
-            Method m = new Method(this.RuntimeContractType, null, ContractNodes.ReportFailureName, pl, SystemTypes.Void, body);
-
-            m.Flags = MethodFlags.Assembly | MethodFlags.Static;
-            m.Attributes = new AttributeList();
-            
-            this.RuntimeContractType.Members.Add(m);
-
-            Block returnBlock = new Block(new StatementList(new Return()));
-
-            //
-            // Generate the following
-            //
-
-            // // message == null means: yes we handled it. Otherwise it is the localized failure message
-            // var message = RaiseFailureEvent(kind, userMessage, conditionString, inner);
-            // if (message == null) return;
-
-            var messageLocal = new Local(Identifier.For("msg"), SystemTypes.String);
-
-            ExpressionList elist = new ExpressionList();
-            elist.Add(kindParameter);
-            elist.Add(messageParameter);
-            elist.Add(conditionTextParameter);
-            elist.Add(exceptionParameter);
-
-            body.Statements.Add(new AssignmentStatement(messageLocal,
-                new MethodCall(new MemberBinding(null, this.RaiseFailureEventMethod), elist)));
-
-            body.Statements.Add(new Branch(new UnaryExpression(messageLocal, NodeType.LogicalNot), returnBlock));
-
-            body.Statements.Add(
-                new ExpressionStatement(new MethodCall(new MemberBinding(null, this.TriggerFailureMethod),
-                    new ExpressionList(kindParameter, messageLocal, messageParameter, conditionTextParameter,
-                        exceptionParameter))));
-
-            body.Statements.Add(returnBlock);
-
-            return m;
-        }
-
-        private Method MakeRaiseFailureEventMethod()
-        {
-            Parameter kindParameter = new Parameter(Identifier.For("kind"), contractNodes.ContractFailureKind);
-            Parameter messageParameter = new Parameter(Identifier.For("msg"), SystemTypes.String);
-            Parameter conditionTextParameter = new Parameter(Identifier.For("conditionTxt"), SystemTypes.String);
-            Parameter exceptionParameter = new Parameter(Identifier.For("inner"), SystemTypes.Exception);
-            ParameterList pl = new ParameterList(kindParameter, messageParameter, conditionTextParameter,exceptionParameter);
-
-            Block body = new Block(new StatementList());
-            Method m = new Method(this.RuntimeContractType, null, ContractNodes.RaiseContractFailedEventName, pl,
-                SystemTypes.String, body);
-
-            m.Flags = MethodFlags.Assembly | MethodFlags.Static;
-            m.Attributes = new AttributeList();
-
-            this.RuntimeContractType.Members.Add(m);
-
-            //
-            // Generate the following
-            //
-
-            // return String.Format("{0} failed: {1}: {2}", box(kind), conditionText, message);
-
-            var stringFormat3 = SystemTypes.String.GetMethod(Identifier.For("Format"), SystemTypes.String,
-                SystemTypes.Object, SystemTypes.Object, SystemTypes.Object);
-
-            body.Statements.Add(
-                new Return(new MethodCall(new MemberBinding(null, stringFormat3),
-                    new ExpressionList(new Literal("{0} failed: {1}: {2}"),
-                        new BinaryExpression(kindParameter, new Literal(contractNodes.ContractFailureKind), NodeType.Box),
-                        conditionTextParameter, messageParameter))));
-
-            return m;
-        }
-
-        [SuppressMessage("Microsoft.Contracts", "TestAlwaysEvaluatingToAConstant", Justification = "Static field")]
-        private Method MakeTriggerFailureMethod(bool throwOnFailure)
-        {
-            Parameter kindParameter = new Parameter(Identifier.For("kind"), contractNodes.ContractFailureKind);
-            Parameter messageParameter = new Parameter(Identifier.For("msg"), SystemTypes.String);
-            Parameter userMessageParameter = new Parameter(Identifier.For("userMessage"), SystemTypes.String);
-            Parameter conditionTextParameter = new Parameter(Identifier.For("conditionTxt"), SystemTypes.String);
-            Parameter exceptionParameter = new Parameter(Identifier.For("inner"), SystemTypes.Exception);
-            ParameterList pl = new ParameterList(kindParameter, messageParameter, userMessageParameter, conditionTextParameter, exceptionParameter);
-
-            Block body = new Block(new StatementList());
-            Method m = new Method(this.RuntimeContractType, null, ContractNodes.TriggerFailureName, pl, SystemTypes.Void, body);
-
-            m.Flags = MethodFlags.Assembly | MethodFlags.Static;
-            m.Attributes = new AttributeList();
-
-            this.RuntimeContractType.Members.Add(m);
-
-            //
-            // Generate the following
-            //
-            // #if throwOnFailure
-            // throw new ContractException(kind, message, userMessage, conditionText, inner);
-            // #else
-            // Debug.Assert(false, messageParameter);
-            // #endif
-
-            if (throwOnFailure)
-            {
-                ExpressionList elist = new ExpressionList();
-                elist.Add(kindParameter);
-                elist.Add(messageParameter);
-                elist.Add(userMessageParameter);
-                elist.Add(conditionTextParameter);
-                elist.Add(exceptionParameter);
-                body.Statements.Add(new Throw(new Construct(new MemberBinding(null, this.ContractExceptionCtor), elist)));
-            }
-            else
-            {
-                var sysassem = SystemTypes.SystemDllAssembly;
-                if (sysassem == null)
-                {
-                    throw new ApplicationException("Cannot find System.dll");
-                }
-
-                var debugType = sysassem.GetType(Identifier.For("System.Diagnostics"), Identifier.For("Debug"));
-                if (debugType == null)
-                {
-                    throw new ApplicationException("Cannot find System.Diagnostics.Debug");
-                }
-
-                var assertMethod = debugType.GetMethod(Identifier.For("Assert"), SystemTypes.Boolean, SystemTypes.String);
-                if (assertMethod == null)
-                {
-                    throw new ApplicationException("Cannot find System.Diagnostics.Debug.Assert(bool,string)");
-                }
-
-                ExpressionList elist = new ExpressionList();
-                elist.Add(Literal.False);
-                elist.Add(messageParameter);
-                body.Statements.Add(new ExpressionStatement(new MethodCall(new MemberBinding(null, assertMethod), elist)));
-                body.Statements.Add(new Return());
-            }
-
-            return m;
-        }
-
-        /// <summary>
-        /// Called after assembly visit to avoid visiting the generated runtime contract class, we only add it to the target
-        /// here.
-        /// </summary>
-        internal void Commit()
-        {
-            if (this.runtimeContractType != null)
-            {
-                this.targetAssembly.Types.Add(this.runtimeContractType);
-            }
-        }
-    }
-
     [ContractVerification(false)]
     public sealed class Rewriter : Inspector
     {
@@ -5945,239 +3195,6 @@ namespace Microsoft.Contracts.Foxtrot
             enumType.Members.Add(enumValue);
         }
 
-        /// <summary>
-        /// Inserts source strings for Asserts/Assumes (or removes checks based on level)
-        /// Inserts call-site requires checks based on flags.
-        /// </summary>
-        internal sealed class RewriteAssertAssumeAndCallSiteRequires : StandardVisitor
-        {
-            private Dictionary<Method, bool> methodTable;
-            private RuntimeContractMethods runtimeContracts;
-            private RuntimeContractEmitFlags emitFlags;
-
-            private Rewriter parent;
-            private Method containingMethod;
-            private SourceContext lastStmtContext;
-
-            public RewriteAssertAssumeAndCallSiteRequires(Dictionary<Method, bool> replacementMethods,
-                RuntimeContractMethods runtimeContracts, RuntimeContractEmitFlags emitFlags, Rewriter parent,
-                Method containingMethod)
-            {
-                this.methodTable = replacementMethods;
-                this.runtimeContracts = runtimeContracts;
-                this.emitFlags = emitFlags;
-                this.parent = parent;
-                this.containingMethod = containingMethod;
-            }
-
-            // Requires:
-            //  statement.Expression is MethodCall
-            //  statement.Expression.Callee is MemberBinding
-            //  statement.Expression.Callee.BoundMember is Method
-            //  statement.Expression.Callee.BoundMember == "Assert" or "Assume"
-            //
-            //  replacementMethod.ReturnType == methodToReplace.ReturnType
-            //  && replacementMethod.Parameters.Count == 1
-            //  && methodToReplace.Parameters.Count == 1
-            //  && replacementMethod.Parameters[0].Type == methodToReplace.Parameters[0].Type
-            //
-            private static Statement RewriteContractCall(
-                MethodCall call,
-                ExpressionStatement statement,
-                Method /*!*/ methodToReplace,
-                Method /*?*/ replacementMethod)
-            {
-                Contract.Requires(call != null);
-
-                Debug.Assert(call == statement.Expression as MethodCall);
-
-                MemberBinding mb = (MemberBinding) call.Callee;
-                Debug.Assert(mb.BoundMember == methodToReplace);
-
-                mb.BoundMember = replacementMethod;
-
-                if (call.Operands.Count == 1)
-                {
-                    call.Operands.Add(Literal.Null);
-                }
-
-#if false
-        SourceContext sctx = statement.SourceContext;
-        if (sctx.IsValid && sctx.Document.Text != null && sctx.Document.Text.Source != null)
-        {
-          call.Operands.Add(new Literal(sctx.Document.Text.Source, SystemTypes.String));
-        }
-#else
-                ContractAssumeAssertStatement casas = statement as ContractAssumeAssertStatement;
-                if (casas != null)
-                {
-                    call.Operands.Add(new Literal(casas.SourceText, SystemTypes.String));
-                }
-#endif
-                else
-                {
-                    call.Operands.Add(Literal.Null);
-                    //call.Operands.Add(new Literal("No other information available", SystemTypes.String));
-                }
-
-                return statement;
-            }
-
-            public override StatementList VisitStatementList(StatementList statements)
-            {
-                if (statements == null) return null;
-
-                for (int i = 0; i < statements.Count; i++)
-                {
-                    var stmt = statements[i];
-                    if (stmt == null) continue;
-
-                    if (stmt.SourceContext.IsValid)
-                    {
-                        this.lastStmtContext = stmt.SourceContext;
-                    }
-
-                    statements[i] = (Statement) base.Visit(stmt);
-                }
-
-                return statements;
-            }
-
-            public override Statement VisitExpressionStatement(ExpressionStatement statement)
-            {
-                MethodCall call;
-                Method contractMethod = Rewriter.ExtractCallFromStatement(statement, out call);
-                if (contractMethod == null)
-                {
-                    return base.VisitExpressionStatement(statement);
-                }
-
-                Method keyToLookUp = contractMethod;
-                if (this.methodTable.ContainsKey(keyToLookUp))
-                {
-                    var isAssert = keyToLookUp.Name.Name == "Assert";
-
-                    Method repMethod;
-                    if (isAssert)
-                    {
-                        repMethod = this.runtimeContracts.AssertMethod;
-                    }
-                    else
-                    {
-                        // assume it is Assume
-                        repMethod = this.runtimeContracts.AssumeMethod;
-                    }
-
-                    if (repMethod == null)
-                    {
-                        return base.VisitExpressionStatement(statement);
-                    }
-
-                    if (isAssert &&
-                        (!this.emitFlags.Emit(RuntimeContractEmitFlags.Asserts) ||
-                         this.runtimeContracts.PublicSurfaceOnly))
-                    {
-                        var pops = CountPopExpressions.Count(call);
-                        if (pops > 0)
-                        {
-                            var block = new Block(new StatementList(pops));
-                            while (pops-- > 0)
-                            {
-                                // emit a pop
-                                block.Statements.Add(new ExpressionStatement(new UnaryExpression(null, NodeType.Pop)));
-                            }
-
-                            return block;
-                        }
-
-                        return null;
-                    }
-
-                    if (!isAssert &&
-                        (!this.emitFlags.Emit(RuntimeContractEmitFlags.Assumes) ||
-                         this.runtimeContracts.PublicSurfaceOnly))
-                    {
-                        var pops = CountPopExpressions.Count(call);
-                        if (pops > 0)
-                        {
-                            var block = new Block(new StatementList(pops));
-                            while (pops-- > 0)
-                            {
-                                // emit a pop
-                                block.Statements.Add(new ExpressionStatement(new UnaryExpression(null, NodeType.Pop)));
-                            }
-
-                            return block;
-                        }
-
-                        return null;
-                    }
-
-                    var result = RewriteContractCall(call, statement, contractMethod, repMethod);
-
-                    return result;
-                }
-                
-                return base.VisitExpressionStatement(statement);
-            }
-
-            public override Expression VisitMethodCall(MethodCall call)
-            {
-                // if call-site requires are needed, create wrapper types here:
-                MemberBinding mb = call.Callee as MemberBinding;
-                if (mb != null)
-                {
-                    Method m = mb.BoundMember as Method;
-                    if (m != null)
-                    {
-                        bool virtcall = call.NodeType == NodeType.Callvirt;
-                        Method methodWithContract;
-
-                        if (parent.NeedCallSiteRequires(m, virtcall, out methodWithContract))
-                        {
-                            // materialize generic wrapper type
-                            var wrapped = parent.WrapperMethod(m, virtcall, call.Constraint, containingMethod,
-                                methodWithContract, this.lastStmtContext);
-
-                            if (wrapped != null)
-                            {
-                                if (IsProtected(wrapped))
-                                {
-                                    // wrapper was produced in same declaring type, still an instance call with same parameters
-                                    call.Callee = new MemberBinding(mb.TargetObject, wrapped);
-                                }
-                                else
-                                {
-                                    // wrapper is now static in a wrapper class
-                                    call.Callee = new MemberBinding(null, wrapped);
-                                    call.NodeType = NodeType.Call;
-                                    call.Constraint = null;
-
-                                    if (mb.TargetObject != null)
-                                    {
-                                        // instance into static, need to add target object as first parameter
-                                        var origargcount = call.Operands == null ? 0 : call.Operands.Count;
-                                        var elist = new ExpressionList(origargcount + 1);
-
-                                        elist.Add(mb.TargetObject);
-
-                                        for (int i = 0; i < origargcount; i++)
-                                        {
-                                            elist.Add(call.Operands[i]);
-                                        }
-
-                                        call.Operands = elist;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-
-                return base.VisitMethodCall(call);
-            }
-        }
-
         // Call-site Requires instrumentation
 
         private Dictionary<string, TypeNode> wrapperTypes = new Dictionary<string, TypeNode>();
@@ -6433,7 +3450,7 @@ namespace Microsoft.Contracts.Foxtrot
             return InstantiateWrapperMethod(constraint, wrapperMethod, instanceMethod);
         }
 
-        private static bool IsProtected(Method instanceMethod)
+        internal static bool IsProtected(Method instanceMethod)
         {
             return instanceMethod.IsFamily || instanceMethod.IsFamilyAndAssembly || instanceMethod.IsFamilyOrAssembly;
         }
@@ -6441,8 +3458,7 @@ namespace Microsoft.Contracts.Foxtrot
         /// <summary>
         /// properly instantiate wrapper according to original instance (including declaring type and method instance)
         /// </summary>
-        private Method InstantiateWrapperMethod(TypeNode virtcallConstraint, Method wrapperMethod,
-            Method originalInstanceMethod)
+        private Method InstantiateWrapperMethod(TypeNode virtcallConstraint, Method wrapperMethod, Method originalInstanceMethod)
         {
             // if CCI does it for us, great. Otherwise, we have to fake it:
             var instWrapperType = wrapperMethod.DeclaringType;
@@ -7042,163 +4058,235 @@ namespace Microsoft.Contracts.Foxtrot
     }
 
     /// <summary>
-    /// The Extractor has removed all dependencies on whatever contract class was
-    /// used in the assembly. (See the class ExtractAllContractNodesDependencies.)
-    /// 
-    /// This class exists just to make sure that those pseudo-nodes are replaced with
-    /// things the Writer won't barf on.
-    /// For instance, don't leave any ReturnValue or OldExpression ASTs in the assembly.
+    /// Inserts source strings for Asserts/Assumes (or removes checks based on level)
+    /// Inserts call-site requires checks based on flags.
     /// </summary>
-    internal sealed class CleanUpOldAndResult : StandardVisitor
+    internal sealed class RewriteAssertAssumeAndCallSiteRequires : StandardVisitor
     {
-        public override Expression VisitOldExpression(OldExpression oldExpression)
+        private Dictionary<Method, bool> methodTable;
+        private RuntimeContractMethods runtimeContracts;
+        private RuntimeContractEmitFlags emitFlags;
+
+        private Rewriter parent;
+        private Method containingMethod;
+        private SourceContext lastStmtContext;
+
+        public RewriteAssertAssumeAndCallSiteRequires(Dictionary<Method, bool> replacementMethods,
+            RuntimeContractMethods runtimeContracts, RuntimeContractEmitFlags emitFlags, Rewriter parent,
+            Method containingMethod)
         {
-            //Debug.Assert(false, "old was not substituted");
-            TypeNode returnType = oldExpression.Type;
-            if (returnType.IsValueType)
-                return new Literal(0, returnType);
-            
-            return new Literal(null, returnType);
+            this.methodTable = replacementMethods;
+            this.runtimeContracts = runtimeContracts;
+            this.emitFlags = emitFlags;
+            this.parent = parent;
+            this.containingMethod = containingMethod;
         }
 
-        public override Expression VisitReturnValue(ReturnValue returnValue)
+        // Requires:
+        //  statement.Expression is MethodCall
+        //  statement.Expression.Callee is MemberBinding
+        //  statement.Expression.Callee.BoundMember is Method
+        //  statement.Expression.Callee.BoundMember == "Assert" or "Assume"
+        //
+        //  replacementMethod.ReturnType == methodToReplace.ReturnType
+        //  && replacementMethod.Parameters.Count == 1
+        //  && methodToReplace.Parameters.Count == 1
+        //  && replacementMethod.Parameters[0].Type == methodToReplace.Parameters[0].Type
+        //
+        private static Statement RewriteContractCall(
+            MethodCall call,
+            ExpressionStatement statement,
+            Method /*!*/ methodToReplace,
+            Method /*?*/ replacementMethod)
         {
-            // return a default value of the same type as the return value
-            TypeNode returnType = returnValue.Type;
-            ITypeParameter itp = returnType as ITypeParameter;
-            if (itp != null)
+            Contract.Requires(call != null);
+
+            Debug.Assert(call == statement.Expression as MethodCall);
+
+            MemberBinding mb = (MemberBinding) call.Callee;
+            Debug.Assert(mb.BoundMember == methodToReplace);
+
+            mb.BoundMember = replacementMethod;
+
+            if (call.Operands.Count == 1)
             {
-                Local loc = new Local(returnType);
-
-                UnaryExpression loca = new UnaryExpression(loc, NodeType.AddressOf, loc.Type.GetReferenceType());
-                StatementList statements = new StatementList(2);
-
-                statements.Add(new AssignmentStatement(new AddressDereference(loca, returnType, false, 0),
-                    new Literal(null, SystemTypes.Object)));
-
-                statements.Add(new ExpressionStatement(loc));
-
-                return new BlockExpression(new Block(statements), returnType);
+                call.Operands.Add(Literal.Null);
             }
 
-            if (returnType.IsValueType)
-                return new Literal(0, returnType);
-            
-            return new Literal(null, returnType);
-        }
-    }
-
-    /// <summary>
-    /// There is no need in the rewritten assembly for the contract classes.
-    /// And they just cause problems if there is a disagreement between a
-    /// contract class that is in the real assembly and the one in the reference
-    /// assembly. (Admittedly a low probability event, but it has occurred in our
-    /// hand-generated reference assemblies for the framework.)
-    /// </summary>
-    internal sealed class RemoveContractClasses : Inspector
-    {
-        private static void ScrubAttributeList(AttributeList attributes)
+#if false
+        SourceContext sctx = statement.SourceContext;
+        if (sctx.IsValid && sctx.Document.Text != null && sctx.Document.Text.Source != null)
         {
-            if (attributes == null) return;
-            for (int i = 0, n = attributes.Count; i < n; i++)
+          call.Operands.Add(new Literal(sctx.Document.Text.Source, SystemTypes.String));
+        }
+#else
+            ContractAssumeAssertStatement casas = statement as ContractAssumeAssertStatement;
+            if (casas != null)
             {
-                if (attributes[i] == null) continue;
+                call.Operands.Add(new Literal(casas.SourceText, SystemTypes.String));
+            }
+#endif
+            else
+            {
+                call.Operands.Add(Literal.Null);
+                //call.Operands.Add(new Literal("No other information available", SystemTypes.String));
+            }
 
-                if (attributes[i].Type == null) continue;
+            return statement;
+        }
 
-                if (ContractNodes.ContractClassAttributeName.Matches(attributes[i].Type.Name))
+        public override StatementList VisitStatementList(StatementList statements)
+        {
+            if (statements == null) return null;
+
+            for (int i = 0; i < statements.Count; i++)
+            {
+                var stmt = statements[i];
+                if (stmt == null) continue;
+
+                if (stmt.SourceContext.IsValid)
                 {
-                    attributes[i] = null;
+                    this.lastStmtContext = stmt.SourceContext;
                 }
+
+                statements[i] = (Statement) base.Visit(stmt);
             }
+
+            return statements;
         }
 
-        public override void VisitTypeNode(TypeNode typeNode)
+        public override Statement VisitExpressionStatement(ExpressionStatement statement)
         {
-            if (typeNode == null) return;
-
-            ScrubAttributeList(typeNode.Attributes);
-
-            this.VisitMemberList(typeNode.Members);
-        }
-
-        public override void VisitMemberList(MemberList members)
-        {
-            if (members == null) return;
-
-            for (int i = 0, n = members.Count; i < n; i++)
+            MethodCall call;
+            Method contractMethod = Rewriter.ExtractCallFromStatement(statement, out call);
+            if (contractMethod == null)
             {
-                var type = members[i] as TypeNode;
+                return base.VisitExpressionStatement(statement);
+            }
 
-                if (type == null) continue;
+            Method keyToLookUp = contractMethod;
+            if (this.methodTable.ContainsKey(keyToLookUp))
+            {
+                var isAssert = keyToLookUp.Name.Name == "Assert";
 
-                Class c = type as Class;
-                if (c != null && HelperMethods.IsContractTypeForSomeOtherType(c))
+                Method repMethod;
+                if (isAssert)
                 {
-                    members[i] = null;
+                    repMethod = this.runtimeContracts.AssertMethod;
                 }
                 else
                 {
-                    // for nested types
-                    this.VisitTypeNode(type);
+                    // assume it is Assume
+                    repMethod = this.runtimeContracts.AssumeMethod;
                 }
+
+                if (repMethod == null)
+                {
+                    return base.VisitExpressionStatement(statement);
+                }
+
+                if (isAssert &&
+                    (!this.emitFlags.Emit(RuntimeContractEmitFlags.Asserts) ||
+                     this.runtimeContracts.PublicSurfaceOnly))
+                {
+                    var pops = CountPopExpressions.Count(call);
+                    if (pops > 0)
+                    {
+                        var block = new Block(new StatementList(pops));
+                        while (pops-- > 0)
+                        {
+                            // emit a pop
+                            block.Statements.Add(new ExpressionStatement(new UnaryExpression(null, NodeType.Pop)));
+                        }
+
+                        return block;
+                    }
+
+                    return null;
+                }
+
+                if (!isAssert &&
+                    (!this.emitFlags.Emit(RuntimeContractEmitFlags.Assumes) ||
+                     this.runtimeContracts.PublicSurfaceOnly))
+                {
+                    var pops = CountPopExpressions.Count(call);
+                    if (pops > 0)
+                    {
+                        var block = new Block(new StatementList(pops));
+                        while (pops-- > 0)
+                        {
+                            // emit a pop
+                            block.Statements.Add(new ExpressionStatement(new UnaryExpression(null, NodeType.Pop)));
+                        }
+
+                        return block;
+                    }
+
+                    return null;
+                }
+
+                var result = RewriteContractCall(call, statement, contractMethod, repMethod);
+
+                return result;
             }
+                
+            return base.VisitExpressionStatement(statement);
         }
 
-        public override void VisitTypeNodeList(TypeNodeList types)
+        public override Expression VisitMethodCall(MethodCall call)
         {
-            if (types == null) return;
-
-            for (int i = 0, n = types.Count; i < n; i++)
+            // if call-site requires are needed, create wrapper types here:
+            MemberBinding mb = call.Callee as MemberBinding;
+            if (mb != null)
             {
-                var type = types[i];
-                if (type == null) continue;
+                Method m = mb.BoundMember as Method;
+                if (m != null)
+                {
+                    bool virtcall = call.NodeType == NodeType.Callvirt;
+                    Method methodWithContract;
 
-                Class c = type as Class;
-                if (c != null && HelperMethods.IsContractTypeForSomeOtherType(c))
-                {
-                    types[i] = null;
-                }
-                else
-                {
-                    // for nested types
-                    this.VisitTypeNode(type);
+                    if (parent.NeedCallSiteRequires(m, virtcall, out methodWithContract))
+                    {
+                        // materialize generic wrapper type
+                        var wrapped = parent.WrapperMethod(m, virtcall, call.Constraint, containingMethod,
+                            methodWithContract, this.lastStmtContext);
+
+                        if (wrapped != null)
+                        {
+                            if (Rewriter.IsProtected(wrapped))
+                            {
+                                // wrapper was produced in same declaring type, still an instance call with same parameters
+                                call.Callee = new MemberBinding(mb.TargetObject, wrapped);
+                            }
+                            else
+                            {
+                                // wrapper is now static in a wrapper class
+                                call.Callee = new MemberBinding(null, wrapped);
+                                call.NodeType = NodeType.Call;
+                                call.Constraint = null;
+
+                                if (mb.TargetObject != null)
+                                {
+                                    // instance into static, need to add target object as first parameter
+                                    var origargcount = call.Operands == null ? 0 : call.Operands.Count;
+                                    var elist = new ExpressionList(origargcount + 1);
+
+                                    elist.Add(mb.TargetObject);
+
+                                    for (int i = 0; i < origargcount; i++)
+                                    {
+                                        elist.Add(call.Operands[i]);
+                                    }
+
+                                    call.Operands = elist;
+                                }
+                            }
+                        }
+                    }
                 }
             }
-        }
-    }
 
-    /// <summary>
-    /// Application-level exceptions thrown by this component.
-    /// </summary>
-    public class RewriteException : Exception
-    {
-        /// <summary>
-        /// Exception specific to an error occurring in the contract rewriter
-        /// </summary>
-        public RewriteException()
-        {
-        }
-
-        /// <summary>
-        /// Exception specific to an error occurring in the contract rewriter
-        /// </summary>
-        public RewriteException(string s) : base(s)
-        {
-        }
-
-        /// <summary>
-        /// Exception specific to an error occurring in the contract rewriter
-        /// </summary>
-        public RewriteException(string s, Exception inner) : base(s, inner)
-        {
-        }
-
-        /// <summary>
-        /// Exception specific to an error occurring in the contract rewriter
-        /// </summary>
-        public RewriteException(SerializationInfo info, StreamingContext context) : base(info, context)
-        {
+            return base.VisitMethodCall(call);
         }
     }
 }
